Get PGS Catalog Publications

Retrieves PGS publications via the PGS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all publications that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the associations that match simultaneously all criteria provided, then set set_operation to 'intersection'.

get_publications(
  pgp_id = NULL,
  pgs_id = NULL,
  pubmed_id = NULL,
  author = NULL,
  set_operation = "union",
  interactive = TRUE,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

pgp_id

A character vector of PGS Catalog publication accession identifiers.

pgs_id

A character vector of PGS Catalog score accession identifiers.

pubmed_id

An integer vector of PubMed identifiers.

author

A character vector of author names, any author in the list of authors in a publication, .e.g. 'Mavaddat'.

set_operation

Either 'union' or 'intersection'. This tells how publications retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same publications found with different criteria.

interactive

A logical. If all publications are requested, whether to ask interactively if we really want to proceed.

verbose

A logical indicating whether the function should be verbose about the different queries or not.

warnings

A logical indicating whether to print warnings, if any.

progress_bar

Whether to show a progress bar as the queries are performed.

Value

A publications object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search. For more details see the help vignette: vignette("getting-pgs-publications", package = "quincunx").

Examples

# Get PGS publications by their identifier
get_publications(pgp_id = c('PGP000001', 'PGP000002'))
#> An object of class "publications"
#> Slot "publications":
#> # A tibble: 2 × 8
#>   pgp_id    pubmed_id publication_date publication   title autho…¹ doi   authors
#>   <chr>     <chr>     <date>           <chr>         <chr> <chr>   <chr> <chr>  
#> 1 PGP000001 25855707  2015-04-08       J Natl Cance… Pred… Mavadd… 10.1… Mavadd…
#> 2 PGP000002 30554720  2018-12-13       Am J Hum Gen… Poly… Mavadd… 10.1… Mavadd…
#> # … with abbreviated variable name ¹​author_fullname
#> 
#> Slot "pgs_ids":
#> # A tibble: 21 × 3
#>    pgp_id    pgs_id    stage   
#>    <chr>     <chr>     <chr>   
#>  1 PGP000001 PGS000001 gwas/dev
#>  2 PGP000001 PGS000002 gwas/dev
#>  3 PGP000001 PGS000003 gwas/dev
#>  4 PGP000001 PGS000001 eval    
#>  5 PGP000001 PGS000002 eval    
#>  6 PGP000001 PGS000003 eval    
#>  7 PGP000002 PGS000004 gwas/dev
#>  8 PGP000002 PGS000005 gwas/dev
#>  9 PGP000002 PGS000006 gwas/dev
#> 10 PGP000002 PGS000007 gwas/dev
#> # … with 11 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

# By polygenic score identifier
get_publications(pgs_id = 'PGS000003')
#> An object of class "publications"
#> Slot "publications":
#> # A tibble: 1 × 8
#>   pgp_id    pubmed_id publication_date publication   title autho…¹ doi   authors
#>   <chr>     <chr>     <date>           <chr>         <chr> <chr>   <chr> <chr>  
#> 1 PGP000001 25855707  2015-04-08       J Natl Cance… Pred… Mavadd… 10.1… Mavadd…
#> # … with abbreviated variable name ¹​author_fullname
#> 
#> Slot "pgs_ids":
#> # A tibble: 6 × 3
#>   pgp_id    pgs_id    stage   
#>   <chr>     <chr>     <chr>   
#> 1 PGP000001 PGS000001 gwas/dev
#> 2 PGP000001 PGS000002 gwas/dev
#> 3 PGP000001 PGS000003 gwas/dev
#> 4 PGP000001 PGS000001 eval    
#> 5 PGP000001 PGS000002 eval    
#> 6 PGP000001 PGS000003 eval    
#> 

# By PubMed identifier
get_publications(pubmed_id = '30554720')
#> An object of class "publications"
#> Slot "publications":
#> # A tibble: 1 × 8
#>   pgp_id    pubmed_id publication_date publication   title autho…¹ doi   authors
#>   <chr>     <chr>     <date>           <chr>         <chr> <chr>   <chr> <chr>  
#> 1 PGP000002 30554720  2018-12-13       Am J Hum Gen… Poly… Mavadd… 10.1… Mavadd…
#> # … with abbreviated variable name ¹​author_fullname
#> 
#> Slot "pgs_ids":
#> # A tibble: 15 × 3
#>    pgp_id    pgs_id    stage   
#>    <chr>     <chr>     <chr>   
#>  1 PGP000002 PGS000004 gwas/dev
#>  2 PGP000002 PGS000005 gwas/dev
#>  3 PGP000002 PGS000006 gwas/dev
#>  4 PGP000002 PGS000007 gwas/dev
#>  5 PGP000002 PGS000008 gwas/dev
#>  6 PGP000002 PGS000009 gwas/dev
#>  7 PGP000002 PGS000001 eval    
#>  8 PGP000002 PGS000002 eval    
#>  9 PGP000002 PGS000003 eval    
#> 10 PGP000002 PGS000004 eval    
#> 11 PGP000002 PGS000005 eval    
#> 12 PGP000002 PGS000006 eval    
#> 13 PGP000002 PGS000007 eval    
#> 14 PGP000002 PGS000008 eval    
#> 15 PGP000002 PGS000009 eval    
#> 

# By author's last name
get_publications(author = 'Natarajan')
#> An object of class "publications"
#> Slot "publications":
#> # A tibble: 12 × 8
#>    pgp_id    pubmed_id publication_date publication  title autho…¹ doi   authors
#>    <chr>     <chr>     <date>           <chr>        <chr> <chr>   <chr> <chr>  
#>  1 PGP000006 30104762  2018-08-13       Nat Genet    Geno… Khera … 10.1… Khera …
#>  2 PGP000030 31676865  2019-11-01       Nat Genet    Geno… Klarin… 10.1… Klarin…
#>  3 PGP000042 28223407  2017-02-21       Circulation  Poly… Natara… 10.1… Natara…
#>  4 PGP000060 30586733  2019-03-01       Circulation  Whol… Khera … 10.1… Khera …
#>  5 PGP000076 27959714  2016-11-13       N Engl J Med Gene… Khera … 10.1… Khera …
#>  6 PGP000116 32498804  2020-06-01       J Am Coll C… Limi… Aragam… 10.1… Aragam…
#>  7 PGP000127 33021622  2020-10-06       JAMA Cardiol Clin… Trinde… 10.1… Trinde…
#>  8 PGP000159 32981348  2020-09-28       Circulation  Gene… Klarin… 10.1… Klarin…
#>  9 PGP000230 34887591  2021-12-09       Nature       The … Graham… http… Graham…
#> 10 PGP000252 33433237  2021-01-12       Circ Genom … Inte… Ye Y    10.1… Ye Y, …
#> 11 PGP000299 35149155  2022-02-08       Ophthalmolo… Phot… Zekava… 10.1… Zekava…
#> 12 PGP000307 35278270  2022-03-12       Eur J Heart… Gene… Forres… 10.1… Forres…
#> # … with abbreviated variable name ¹​author_fullname
#> 
#> Slot "pgs_ids":
#> # A tibble: 57 × 3
#>    pgp_id    pgs_id    stage   
#>    <chr>     <chr>     <chr>   
#>  1 PGP000006 PGS000013 gwas/dev
#>  2 PGP000006 PGS000014 gwas/dev
#>  3 PGP000006 PGS000015 gwas/dev
#>  4 PGP000006 PGS000016 gwas/dev
#>  5 PGP000006 PGS000017 gwas/dev
#>  6 PGP000006 PGS000013 eval    
#>  7 PGP000006 PGS000014 eval    
#>  8 PGP000006 PGS000015 eval    
#>  9 PGP000006 PGS000016 eval    
#> 10 PGP000006 PGS000017 eval    
#> # … with 47 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>