Get PGS Catalog Releases

This function retrieves PGS Catalog release information. Note that the columns pgs_id, ppm_id and pgp_id contain in each element a vector. These columns can be unnested using unnest_longer (see Examples).

get_releases(
  date = "latest",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

date

One or more dates formatted as "YYYY-MM-DD" for respective releases, "latest" for the latest release, or "all" for all releases.

verbose

Whether to print information about the underlying requests to the REST API server.

warnings

Whether to print warnings about the underlying requests to the REST API server.

progress_bar

Whether to show a progress bar indicating download progress from the REST API server.

Value

A data frame where each row is a release. Columns are:

date

Release date.

n_pgs

Number of released Polygenic Score (PGS) identifiers (pgs_id).

n_ppm

Number of released Performance Metric (PPM) identifiers (ppm_id).

n_pgp

Number of released PGS Catalog Publication (PGP) identifiers (pgp_id).

pgs_id

Released Polygenic Score (PGS) identifiers.

ppm_id

Released Performance Metric (PPM) identifiers.

pgp_id

Released PGS Catalog Publication (PGP) identifiers.

notes

News about the release.

Examples

# Get the latest release
get_releases()
#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 × 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2022-08-03     7    77    12 This release contains 7 new Score(s), 12 new Pub…
#> 
#> Slot "pgs_ids":
#> # A tibble: 7 × 2
#>   date       pgs_id   
#>   <date>     <chr>    
#> 1 2022-08-03 PGS002738
#> 2 2022-08-03 PGS002739
#> 3 2022-08-03 PGS002740
#> 4 2022-08-03 PGS002741
#> 5 2022-08-03 PGS002742
#> 6 2022-08-03 PGS002743
#> 7 2022-08-03 PGS002744
#> 
#> Slot "ppm_ids":
#> # A tibble: 77 × 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2022-08-03 PPM014838
#>  2 2022-08-03 PPM014839
#>  3 2022-08-03 PPM014840
#>  4 2022-08-03 PPM014841
#>  5 2022-08-03 PPM014842
#>  6 2022-08-03 PPM014843
#>  7 2022-08-03 PPM014844
#>  8 2022-08-03 PPM014845
#>  9 2022-08-03 PPM014846
#> 10 2022-08-03 PPM014847
#> # … with 67 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 12 × 2
#>    date       pgp_id   
#>    <date>     <chr>    
#>  1 2022-08-03 PGP000345
#>  2 2022-08-03 PGP000346
#>  3 2022-08-03 PGP000347
#>  4 2022-08-03 PGP000348
#>  5 2022-08-03 PGP000349
#>  6 2022-08-03 PGP000350
#>  7 2022-08-03 PGP000351
#>  8 2022-08-03 PGP000352
#>  9 2022-08-03 PGP000353
#> 10 2022-08-03 PGP000354
#> 11 2022-08-03 PGP000355
#> 12 2022-08-03 PGP000356
#> 
get_releases(date = 'latest')
#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 × 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2022-08-03     7    77    12 This release contains 7 new Score(s), 12 new Pub…
#> 
#> Slot "pgs_ids":
#> # A tibble: 7 × 2
#>   date       pgs_id   
#>   <date>     <chr>    
#> 1 2022-08-03 PGS002738
#> 2 2022-08-03 PGS002739
#> 3 2022-08-03 PGS002740
#> 4 2022-08-03 PGS002741
#> 5 2022-08-03 PGS002742
#> 6 2022-08-03 PGS002743
#> 7 2022-08-03 PGS002744
#> 
#> Slot "ppm_ids":
#> # A tibble: 77 × 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2022-08-03 PPM014838
#>  2 2022-08-03 PPM014839
#>  3 2022-08-03 PPM014840
#>  4 2022-08-03 PPM014841
#>  5 2022-08-03 PPM014842
#>  6 2022-08-03 PPM014843
#>  7 2022-08-03 PPM014844
#>  8 2022-08-03 PPM014845
#>  9 2022-08-03 PPM014846
#> 10 2022-08-03 PPM014847
#> # … with 67 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 12 × 2
#>    date       pgp_id   
#>    <date>     <chr>    
#>  1 2022-08-03 PGP000345
#>  2 2022-08-03 PGP000346
#>  3 2022-08-03 PGP000347
#>  4 2022-08-03 PGP000348
#>  5 2022-08-03 PGP000349
#>  6 2022-08-03 PGP000350
#>  7 2022-08-03 PGP000351
#>  8 2022-08-03 PGP000352
#>  9 2022-08-03 PGP000353
#> 10 2022-08-03 PGP000354
#> 11 2022-08-03 PGP000355
#> 12 2022-08-03 PGP000356
#> 

# Get all releases
get_releases(date = 'all')
#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 51 × 5
#>    date       n_pgs n_ppm n_pgp notes                                           
#>    <date>     <int> <int> <int> <chr>                                           
#>  1 2022-08-03     7    77    12 This release contains 7 new Score(s), 12 new Pu…
#>  2 2022-07-21     8    33     6 This release contains 8 new Score(s), 6 new Pub…
#>  3 2022-07-05     1    10     1 This release contains 1 new Score(s), 1 new Pub…
#>  4 2022-06-29     8    73     8 This release contains 8 new Score(s), 8 new Pub…
#>  5 2022-06-09   429  1701     5 This release contains 429 new Score(s), 5 new P…
#>  6 2022-05-18     9    40    13 This release contains 9 new Score(s), 13 new Pu…
#>  7 2022-04-13    13    75    10 This release contains 13 new Score(s), 10 new P…
#>  8 2022-04-01     9    38     8 This release contains 9 new Score(s), 8 new Pub…
#>  9 2022-03-16     8    48    11 This release contains 8 new Score(s), 11 new Pu…
#> 10 2022-02-23     3    15     2 This release contains 3 new Score(s), 2 new Pub…
#> # … with 41 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "pgs_ids":
#> # A tibble: 2,652 × 2
#>    date       pgs_id   
#>    <date>     <chr>    
#>  1 2022-08-03 PGS002738
#>  2 2022-08-03 PGS002739
#>  3 2022-08-03 PGS002740
#>  4 2022-08-03 PGS002741
#>  5 2022-08-03 PGS002742
#>  6 2022-08-03 PGS002743
#>  7 2022-08-03 PGS002744
#>  8 2022-07-21 PGS002238
#>  9 2022-07-21 PGS002239
#> 10 2022-07-21 PGS002732
#> # … with 2,642 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "ppm_ids":
#> # A tibble: 12,608 × 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2022-08-03 PPM014838
#>  2 2022-08-03 PPM014839
#>  3 2022-08-03 PPM014840
#>  4 2022-08-03 PPM014841
#>  5 2022-08-03 PPM014842
#>  6 2022-08-03 PPM014843
#>  7 2022-08-03 PPM014844
#>  8 2022-08-03 PPM014845
#>  9 2022-08-03 PPM014846
#> 10 2022-08-03 PPM014847
#> # … with 12,598 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 350 × 2
#>    date       pgp_id   
#>    <date>     <chr>    
#>  1 2022-08-03 PGP000345
#>  2 2022-08-03 PGP000346
#>  3 2022-08-03 PGP000347
#>  4 2022-08-03 PGP000348
#>  5 2022-08-03 PGP000349
#>  6 2022-08-03 PGP000350
#>  7 2022-08-03 PGP000351
#>  8 2022-08-03 PGP000352
#>  9 2022-08-03 PGP000353
#> 10 2022-08-03 PGP000354
#> # … with 340 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

# Get a specific release by date
get_releases(date = '2020-08-19')
#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 × 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2020-08-19    22    67     5 This release contains 22 new Score(s), 5 new Pub…
#> 
#> Slot "pgs_ids":
#> # A tibble: 22 × 2
#>    date       pgs_id   
#>    <date>     <chr>    
#>  1 2020-08-19 PGS000296
#>  2 2020-08-19 PGS000297
#>  3 2020-08-19 PGS000298
#>  4 2020-08-19 PGS000299
#>  5 2020-08-19 PGS000300
#>  6 2020-08-19 PGS000301
#>  7 2020-08-19 PGS000302
#>  8 2020-08-19 PGS000303
#>  9 2020-08-19 PGS000304
#> 10 2020-08-19 PGS000305
#> # … with 12 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "ppm_ids":
#> # A tibble: 67 × 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2020-08-19 PPM000743
#>  2 2020-08-19 PPM000744
#>  3 2020-08-19 PPM000745
#>  4 2020-08-19 PPM000746
#>  5 2020-08-19 PPM000747
#>  6 2020-08-19 PPM000748
#>  7 2020-08-19 PPM000749
#>  8 2020-08-19 PPM000750
#>  9 2020-08-19 PPM000751
#> 10 2020-08-19 PPM000752
#> # … with 57 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 5 × 2
#>   date       pgp_id   
#>   <date>     <chr>    
#> 1 2020-08-19 PGP000090
#> 2 2020-08-19 PGP000091
#> 3 2020-08-19 PGP000092
#> 4 2020-08-19 PGP000093
#> 5 2020-08-19 PGP000094
#>