Get PGS Catalog Sample Sets

Retrieves sample sets via the PGS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all sample sets that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the associations that match simultaneously all criteria provided, then set set_operation to 'intersection'.

get_sample_sets(
  pss_id = NULL,
  pgs_id = NULL,
  set_operation = "union",
  interactive = TRUE,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

pss_id

A character vector of PGS Catalog sample sets accession identifiers.

pgs_id

A character vector of PGS Catalog score accession identifiers.

set_operation

Either 'union' or 'intersection'. This tells how performance metrics retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same sample sets found with different criteria.

interactive

A logical. If all sample sets are requested, whether to ask interactively if we really want to proceed.

verbose

A logical indicating whether the function should be verbose about the different queries or not.

warnings

A logical indicating whether to print warnings, if any.

progress_bar

Whether to show a progress bar indicating download progress from the REST API server.

Value

A sample_sets object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search.

Examples

# Search by PGS identifier
get_sample_sets(pgs_id = 'PGS000013')
#> An object of class "sample_sets"
#> Slot "sample_sets":
#> # A tibble: 45 × 1
#>    pss_id   
#>    <chr>    
#>  1 PSS000015
#>  2 PSS000019
#>  3 PSS000020
#>  4 PSS000021
#>  5 PSS000022
#>  6 PSS000219
#>  7 PSS000227
#>  8 PSS000228
#>  9 PSS000229
#> 10 PSS000230
#> # … with 35 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "samples":
#> # A tibble: 49 × 15
#>    pss_id  sampl…¹ stage sampl…² sampl…³ sampl…⁴ sampl…⁵ pheno…⁶ ances…⁷ ances…⁸
#>    <chr>     <int> <chr>   <int>   <int>   <int>   <dbl> <chr>   <chr>   <chr>  
#>  1 PSS000…       1 eval   288978    8676  280302    NA   CAD as… Europe… NA     
#>  2 PSS000…       1 eval     5762     173    5589    41.3 Preval… Europe… French…
#>  3 PSS000…       1 eval      862     446     416    NA   Recurr… Europe… French…
#>  4 PSS000…       2 eval     2333     937    1396    NA   Recurr… Europe… French…
#>  5 PSS000…       1 eval     1964     974     976    72.7 Preval… Europe… French…
#>  6 PSS000…       1 eval     3309    2492     817    72.4 Preval… Europe… French…
#>  7 PSS000…       1 eval    11010     126   10884    17.1 Phenot… Europe… NA     
#>  8 PSS000…       1 eval      544      40     504    NA   NA      Asian … NA     
#>  9 PSS000…       1 eval     1298     336     962    NA   NA      Africa… NA     
#> 10 PSS000…       1 eval      919     168     751    NA   NA      Hispan… NA     
#> # … with 39 more rows, 5 more variables: country <chr>,
#> #   ancestry_additional_description <chr>, study_id <chr>, pubmed_id <chr>,
#> #   cohorts_additional_description <chr>, and abbreviated variable names
#> #   ¹​sample_id, ²​sample_size, ³​sample_cases, ⁴​sample_controls,
#> #   ⁵​sample_percent_male, ⁶​phenotype_description, ⁷​ancestry_category, ⁸​ancestry
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> Slot "demographics":
#> # A tibble: 16 × 11
#>    pss_id  sampl…¹ varia…² estim…³ estim…⁴ unit  varia…⁵ varia…⁶ inter…⁷ inter…⁸
#>    <chr>     <int> <chr>   <chr>     <dbl> <chr> <chr>     <dbl> <chr>     <dbl>
#>  1 PSS000…       1 age     mean       34   years NA           NA iqr        30  
#>  2 PSS000…       2 age     mean       33   years NA           NA iqr        30  
#>  3 PSS000…       1 age     mean       54   years NA           NA iqr        46  
#>  4 PSS000…       2 age     mean       55   years NA           NA iqr        49  
#>  5 PSS000…       1 age     mean       60.6 years NA           NA iqr        54.4
#>  6 PSS000…       2 age     mean       52.8 years NA           NA iqr        46.3
#>  7 PSS000…       1 follow… median      9.2 years NA           NA iqr         5.5
#>  8 PSS000…       1 follow… median      9.2 years NA           NA iqr         5.5
#>  9 PSS000…       1 follow… median     11.7 years NA           NA iqr         6  
#> 10 PSS000…       1 follow… median     11.7 years NA           NA iqr         6  
#> 11 PSS000…       1 follow… median     10.4 years NA           NA iqr         5.7
#> 12 PSS000…       1 follow… median     10.4 years NA           NA iqr         5.7
#> 13 PSS000…       1 follow… median     21.3 years NA           NA iqr        16.1
#> 14 PSS000…       1 follow… median     23.2 years NA           NA iqr        17.6
#> 15 PSS000…       1 follow… median      8.1 years NA           NA iqr         7.4
#> 16 PSS001…       1 follow… median     14   years NA           NA iqr        14  
#> # … with 1 more variable: interval_upper <dbl>, and abbreviated variable names
#> #   ¹​sample_id, ²​variable, ³​estimate_type, ⁴​estimate, ⁵​variability_type,
#> #   ⁶​variability, ⁷​interval_type, ⁸​interval_lower
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "cohorts":
#> # A tibble: 114 × 4
#>    pss_id    sample_id cohort_symbol cohort_name                                
#>    <chr>         <int> <chr>         <chr>                                      
#>  1 PSS000015         1 UKB           UK Biobank                                 
#>  2 PSS000019         1 CARTaGENE     CARTaGENE cohort (CHU Sainte-Justine, Queb…
#>  3 PSS000020         1 MHI           Montreal Heart Institute Biobank           
#>  4 PSS000020         2 MHI           Montreal Heart Institute Biobank           
#>  5 PSS000021         1 MHI           Montreal Heart Institute Biobank           
#>  6 PSS000022         1 MHI           Montreal Heart Institute Biobank           
#>  7 PSS000219         1 CG            Color Genomics                             
#>  8 PSS000227         1 VIRGO         Variation in Recovery: Role of Gender on O…
#>  9 PSS000227         1 MESA          Multi-Ethnic Study of Atherosclerosis      
#> 10 PSS000228         1 VIRGO         Variation in Recovery: Role of Gender on O…
#> # … with 104 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

# Search by the PSS identifier
get_sample_sets(pss_id = 'PSS000068')
#> An object of class "sample_sets"
#> Slot "sample_sets":
#> # A tibble: 1 × 1
#>   pss_id   
#>   <chr>    
#> 1 PSS000068
#> 
#> Slot "samples":
#> # A tibble: 3 × 15
#>   pss_id   sampl…¹ stage sampl…² sampl…³ sampl…⁴ sampl…⁵ pheno…⁶ ances…⁷ ances…⁸
#>   <chr>      <int> <chr>   <int>   <int>   <int>   <dbl> <chr>   <chr>   <chr>  
#> 1 PSS0000…       1 eval     3157    3157       0     100 Prosta… Europe… NA     
#> 2 PSS0000…       2 eval     8796    2148    6648     100 Prosta… Europe… NA     
#> 3 PSS0000…       3 eval     5059    4099     960     100 Prosta… Europe… NA     
#> # … with 5 more variables: country <chr>,
#> #   ancestry_additional_description <chr>, study_id <chr>, pubmed_id <chr>,
#> #   cohorts_additional_description <chr>, and abbreviated variable names
#> #   ¹​sample_id, ²​sample_size, ³​sample_cases, ⁴​sample_controls,
#> #   ⁵​sample_percent_male, ⁶​phenotype_description, ⁷​ancestry_category, ⁸​ancestry
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "demographics":
#> # A tibble: 3 × 11
#>   pss_id   sampl…¹ varia…² estim…³ estim…⁴ unit  varia…⁵ varia…⁶ inter…⁷ inter…⁸
#>   <chr>      <int> <chr>   <chr>     <dbl> <chr> <chr>     <dbl> <chr>     <dbl>
#> 1 PSS0000…       1 age     NA           NA years NA           NA range        50
#> 2 PSS0000…       2 age     NA           NA years NA           NA range        50
#> 3 PSS0000…       3 age     NA           NA years NA           NA range        50
#> # … with 1 more variable: interval_upper <dbl>, and abbreviated variable names
#> #   ¹​sample_id, ²​variable, ³​estimate_type, ⁴​estimate, ⁵​variability_type,
#> #   ⁶​variability, ⁷​interval_type, ⁸​interval_lower
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "cohorts":
#> # A tibble: 3 × 4
#>   pss_id    sample_id cohort_symbol cohort_name                                 
#>   <chr>         <int> <chr>         <chr>                                       
#> 1 PSS000068         1 UKGPCS        UK Genetic Prostate Cancer Study            
#> 2 PSS000068         2 ProtecT       Prostate testing for cancer and Treatment s…
#> 3 PSS000068         3 SEARCH        Study of Epidemiology and Risk Factors in C…
#>