Get PGS Catalog Scores

Retrieves polygenic scores via the PGS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all scores that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the associations that match simultaneously all criteria provided, then set set_operation to 'intersection'.

get_scores(
  pgs_id = NULL,
  efo_id = NULL,
  pubmed_id = NULL,
  set_operation = "union",
  interactive = TRUE,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

pgs_id

A character vector of PGS Catalog score accession identifiers.

efo_id

A character vector of EFO identifiers.

pubmed_id

An integer vector of PubMed identifiers.

set_operation

Either 'union' or 'intersection'. This tells how scores retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same scores found with different criteria.

interactive

A logical. If all scores are requested, whether to ask interactively if we really want to proceed.

verbose

A logical indicating whether the function should be verbose about the different queries or not.

warnings

A logical indicating whether to print warnings, if any.

progress_bar

Whether to show a progress bar as the queries are performed.

Value

A scores object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search.

Examples

# By `pgs_id`
get_scores(pgs_id = 'PGS000088')
#> An object of class "scores"
#> Slot "scores":
#> # A tibble: 1 × 12
#>   pgs_id pgs_n…¹ scori…² match…³ repor…⁴ trait…⁵ pgs_m…⁶ pgs_m…⁷ n_var…⁸ n_var…⁹
#>   <chr>  <chr>   <chr>   <lgl>   <chr>   <chr>   <chr>   <chr>     <int>   <int>
#> 1 PGS00… baso    https:… TRUE    Basoph… BASO#   Regula… Bi-all…    9121       0
#> # … with 2 more variables: assembly <chr>, license <chr>, and abbreviated
#> #   variable names ¹​pgs_name, ²​scoring_file, ³​matches_publication,
#> #   ⁴​reported_trait, ⁵​trait_additional_description, ⁶​pgs_method_name,
#> #   ⁷​pgs_method_params, ⁸​n_variants, ⁹​n_variants_interactions
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "publications":
#> # A tibble: 1 × 8
#>   pgs_id    pgp_id    pubmed_id publication_date publication title autho…¹ doi  
#>   <chr>     <chr>     <chr>     <date>           <chr>       <chr> <chr>   <chr>
#> 1 PGS000088 PGP000051 35072137  2022-01-12       Cell Genom  Mach… Xu Y    10.1…
#> # … with abbreviated variable name ¹​author_fullname
#> 
#> Slot "samples":
#> # A tibble: 2 × 15
#>   pgs_id   sampl…¹ stage sampl…² sampl…³ sampl…⁴ sampl…⁵ pheno…⁶ ances…⁷ ances…⁸
#>   <chr>      <int> <chr>   <int>   <int>   <int>   <dbl> <chr>   <chr>   <chr>  
#> 1 PGS0000…       1 gwas   404718      NA      NA      46 NA      Europe… NA     
#> 2 PGS0000…       2 dev    323774      NA      NA      46 NA      Europe… NA     
#> # … with 5 more variables: country <chr>,
#> #   ancestry_additional_description <chr>, study_id <chr>, pubmed_id <chr>,
#> #   cohorts_additional_description <chr>, and abbreviated variable names
#> #   ¹​sample_id, ²​sample_size, ³​sample_cases, ⁴​sample_controls,
#> #   ⁵​sample_percent_male, ⁶​phenotype_description, ⁷​ancestry_category, ⁸​ancestry
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "demographics":
#> # A tibble: 2 × 11
#>   pgs_id   sampl…¹ varia…² estim…³ estim…⁴ unit  varia…⁵ varia…⁶ inter…⁷ inter…⁸
#>   <chr>      <int> <chr>   <chr>     <dbl> <chr> <chr>     <dbl> <chr>     <dbl>
#> 1 PGS0000…       1 age     mean       57.2 years NA           NA range      38.9
#> 2 PGS0000…       2 age     mean       57.2 years NA           NA range      38.9
#> # … with 1 more variable: interval_upper <dbl>, and abbreviated variable names
#> #   ¹​sample_id, ²​variable, ³​estimate_type, ⁴​estimate, ⁵​variability_type,
#> #   ⁶​variability, ⁷​interval_type, ⁸​interval_lower
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "cohorts":
#> # A tibble: 2 × 4
#>   pgs_id    sample_id cohort_symbol cohort_name
#>   <chr>         <int> <chr>         <chr>      
#> 1 PGS000088         1 UKB           UK Biobank 
#> 2 PGS000088         2 UKB           UK Biobank 
#> 
#> Slot "traits":
#> # A tibble: 1 × 5
#>   pgs_id    efo_id      trait          description                         url  
#>   <chr>     <chr>       <chr>          <chr>                               <chr>
#> 1 PGS000088 EFO_0005090 basophil count The number of granulocytes (polymo… http…
#> 
#> Slot "stages_tally":
#> # A tibble: 3 × 4
#>   pgs_id    stage sample_size n_sample_sets
#>   <chr>     <chr>       <int>         <int>
#> 1 PGS000088 gwas       404718            NA
#> 2 PGS000088 dev        323774            NA
#> 3 PGS000088 eval           NA             2
#> 
#> Slot "ancestry_frequencies":
#> # A tibble: 3 × 4
#>   pgs_id    stage ancestry_class_symbol frequency
#>   <chr>     <chr> <chr>                     <dbl>
#> 1 PGS000088 gwas  EUR                         100
#> 2 PGS000088 dev   EUR                         100
#> 3 PGS000088 eval  EUR                         100
#> 
#> Slot "multi_ancestry_composition":
#> # A tibble: 0 × 4
#> # … with 4 variables: pgs_id <chr>, stage <chr>,
#> #   multi_ancestry_class_symbol <chr>, ancestry_class_symbol <chr>
#> # ℹ Use `colnames()` to see all variable names
#> 

# By `efo_id`
get_scores(efo_id = 'EFO_0007992')
#> An object of class "scores"
#> Slot "scores":
#> # A tibble: 3 × 12
#>   pgs_id pgs_n…¹ scori…² match…³ repor…⁴ trait…⁵ pgs_m…⁶ pgs_m…⁷ n_var…⁸ n_var…⁹
#>   <chr>  <chr>   <chr>   <lgl>   <chr>   <chr>   <chr>   <chr>     <int>   <int>
#> 1 PGS00… baso_p  https:… TRUE    Basoph… BASO%   Regula… "Bi-al…    5248       0
#> 2 PGS00… baso_p  https:… TRUE    Basoph… BASO%   Genome… "Condi…     150       0
#> 3 PGS00… GBE_IN… https:… TRUE    Basoph… https:… snpnet   NA        3205       0
#> # … with 2 more variables: assembly <chr>, license <chr>, and abbreviated
#> #   variable names ¹​pgs_name, ²​scoring_file, ³​matches_publication,
#> #   ⁴​reported_trait, ⁵​trait_additional_description, ⁶​pgs_method_name,
#> #   ⁷​pgs_method_params, ⁸​n_variants, ⁹​n_variants_interactions
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "publications":
#> # A tibble: 3 × 8
#>   pgs_id    pgp_id    pubmed_id publication_date publication title autho…¹ doi  
#>   <chr>     <chr>     <chr>     <date>           <chr>       <chr> <chr>   <chr>
#> 1 PGS000089 PGP000051 35072137  2022-01-12       Cell Genom  Mach… Xu Y    10.1…
#> 2 PGS000164 PGP000078 32888494  2020-09-01       Cell        The … Vuckov… 10.1…
#> 3 PGS001377 PGP000244 35324888  2022-03-24       PLoS Genet  Sign… Taniga… 10.1…
#> # … with abbreviated variable name ¹​author_fullname
#> 
#> Slot "samples":
#> # A tibble: 4 × 15
#>   pgs_id   sampl…¹ stage sampl…² sampl…³ sampl…⁴ sampl…⁵ pheno…⁶ ances…⁷ ances…⁸
#>   <chr>      <int> <chr>   <int>   <int>   <int>   <dbl> <chr>   <chr>   <chr>  
#> 1 PGS0000…       1 gwas   404532      NA      NA      46 NA      Europe… NA     
#> 2 PGS0000…       2 dev    323626      NA      NA      46 NA      Europe… NA     
#> 3 PGS0001…       1 gwas   408112      NA      NA      NA NA      Europe… NA     
#> 4 PGS0013…       1 dev    261893      NA      NA      NA NA      Europe… NA     
#> # … with 5 more variables: country <chr>,
#> #   ancestry_additional_description <chr>, study_id <chr>, pubmed_id <chr>,
#> #   cohorts_additional_description <chr>, and abbreviated variable names
#> #   ¹​sample_id, ²​sample_size, ³​sample_cases, ⁴​sample_controls,
#> #   ⁵​sample_percent_male, ⁶​phenotype_description, ⁷​ancestry_category, ⁸​ancestry
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "demographics":
#> # A tibble: 2 × 11
#>   pgs_id   sampl…¹ varia…² estim…³ estim…⁴ unit  varia…⁵ varia…⁶ inter…⁷ inter…⁸
#>   <chr>      <int> <chr>   <chr>     <dbl> <chr> <chr>     <dbl> <chr>     <dbl>
#> 1 PGS0000…       1 age     mean       57.2 years NA           NA range      38.9
#> 2 PGS0000…       2 age     mean       57.2 years NA           NA range      38.9
#> # … with 1 more variable: interval_upper <dbl>, and abbreviated variable names
#> #   ¹​sample_id, ²​variable, ³​estimate_type, ⁴​estimate, ⁵​variability_type,
#> #   ⁶​variability, ⁷​interval_type, ⁸​interval_lower
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "cohorts":
#> # A tibble: 4 × 4
#>   pgs_id    sample_id cohort_symbol cohort_name
#>   <chr>         <int> <chr>         <chr>      
#> 1 PGS000089         1 UKB           UK Biobank 
#> 2 PGS000089         2 UKB           UK Biobank 
#> 3 PGS000164         1 UKB           UK Biobank 
#> 4 PGS001377         1 UKB           UK Biobank 
#> 
#> Slot "traits":
#> # A tibble: 3 × 5
#>   pgs_id    efo_id      trait                             description      url  
#>   <chr>     <chr>       <chr>                             <chr>            <chr>
#> 1 PGS000089 EFO_0007992 basophil percentage of leukocytes A calculated me… http…
#> 2 PGS000164 EFO_0007992 basophil percentage of leukocytes A calculated me… http…
#> 3 PGS001377 EFO_0007992 basophil percentage of leukocytes A calculated me… http…
#> 
#> Slot "stages_tally":
#> # A tibble: 7 × 4
#>   pgs_id    stage sample_size n_sample_sets
#>   <chr>     <chr>       <int>         <int>
#> 1 PGS000089 gwas       404532            NA
#> 2 PGS000089 dev        323626            NA
#> 3 PGS000089 eval           NA             2
#> 4 PGS000164 gwas       408112            NA
#> 5 PGS000164 eval           NA             1
#> 6 PGS001377 dev        261893            NA
#> 7 PGS001377 eval           NA             5
#> 
#> Slot "ancestry_frequencies":
#> # A tibble: 10 × 4
#>    pgs_id    stage ancestry_class_symbol frequency
#>    <chr>     <chr> <chr>                     <dbl>
#>  1 PGS000089 gwas  EUR                         100
#>  2 PGS000089 dev   EUR                         100
#>  3 PGS000089 eval  EUR                         100
#>  4 PGS000164 gwas  EUR                         100
#>  5 PGS000164 eval  EUR                         100
#>  6 PGS001377 dev   EUR                         100
#>  7 PGS001377 eval  AFR                          20
#>  8 PGS001377 eval  EAS                          20
#>  9 PGS001377 eval  EUR                          40
#> 10 PGS001377 eval  SAS                          20
#> 
#> Slot "multi_ancestry_composition":
#> # A tibble: 0 × 4
#> # … with 4 variables: pgs_id <chr>, stage <chr>,
#> #   multi_ancestry_class_symbol <chr>, ancestry_class_symbol <chr>
#> # ℹ Use `colnames()` to see all variable names
#> 

# By `pubmed_id`
get_scores(pubmed_id = '25748612')
#> An object of class "scores"
#> Slot "scores":
#> # A tibble: 1 × 12
#>   pgs_id pgs_n…¹ scori…² match…³ repor…⁴ trait…⁵ pgs_m…⁶ pgs_m…⁷ n_var…⁸ n_var…⁹
#>   <chr>  <chr>   <chr>   <lgl>   <chr>   <chr>   <chr>   <chr>     <int>   <int>
#> 1 PGS00… GRS27   https:… TRUE    Corona… NA      Genome… NA           27       0
#> # … with 2 more variables: assembly <chr>, license <chr>, and abbreviated
#> #   variable names ¹​pgs_name, ²​scoring_file, ³​matches_publication,
#> #   ⁴​reported_trait, ⁵​trait_additional_description, ⁶​pgs_method_name,
#> #   ⁷​pgs_method_params, ⁸​n_variants, ⁹​n_variants_interactions
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "publications":
#> # A tibble: 1 × 8
#>   pgs_id    pgp_id    pubmed_id publication_date publication title autho…¹ doi  
#>   <chr>     <chr>     <chr>     <date>           <chr>       <chr> <chr>   <chr>
#> 1 PGS000010 PGP000003 25748612  2015-03-04       Lancet      Gene… Mega JL 10.1…
#> # … with abbreviated variable name ¹​author_fullname
#> 
#> Slot "samples":
#> # A tibble: 1 × 15
#>   pgs_id   sampl…¹ stage sampl…² sampl…³ sampl…⁴ sampl…⁵ pheno…⁶ ances…⁷ ances…⁸
#>   <chr>      <int> <chr>   <int>   <int>   <int>   <dbl> <chr>   <chr>   <chr>  
#> 1 PGS0000…       1 gwas    86995   22233   64762      NA NA      Europe… NA     
#> # … with 5 more variables: country <chr>,
#> #   ancestry_additional_description <chr>, study_id <chr>, pubmed_id <chr>,
#> #   cohorts_additional_description <chr>, and abbreviated variable names
#> #   ¹​sample_id, ²​sample_size, ³​sample_cases, ⁴​sample_controls,
#> #   ⁵​sample_percent_male, ⁶​phenotype_description, ⁷​ancestry_category, ⁸​ancestry
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "demographics":
#> # A tibble: 0 × 11
#> # … with 11 variables: pgs_id <chr>, sample_id <int>, variable <chr>,
#> #   estimate_type <chr>, estimate <dbl>, unit <chr>, variability_type <chr>,
#> #   variability <dbl>, interval_type <chr>, interval_lower <dbl>,
#> #   interval_upper <dbl>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "cohorts":
#> # A tibble: 12 × 4
#>    pgs_id    sample_id cohort_symbol cohort_name                                
#>    <chr>         <int> <chr>         <chr>                                      
#>  1 PGS000010         1 PennCATH      PennCATH                                   
#>  2 PGS000010         1 WTCCC         Wellcome Trust Case Control Consortium     
#>  3 PGS000010         1 ADVANCE       Atherosclerotic Disease, Vascular function…
#>  4 PGS000010         1 AtheroRemo    AtheroRemo                                 
#>  5 PGS000010         1 CADomics      Coronary Artery Disease and genomics       
#>  6 PGS000010         1 CHARGE        CHARGE                                     
#>  7 PGS000010         1 deCODE        deCODE Genetics                            
#>  8 PGS000010         1 GerMIFS       German Myocardial Infarction Family Study  
#>  9 PGS000010         1 LURIC         Ludwigshafen Risk and Cardiovascular Health
#> 10 PGS000010         1 MedSTAR       MedSTAR cardiac catheterization study      
#> 11 PGS000010         1 MIGen         Myocardial Infarction Genetics Consortium  
#> 12 PGS000010         1 OHGS          Ottawa Heart Genomics Study                
#> 
#> Slot "traits":
#> # A tibble: 1 × 5
#>   pgs_id    efo_id      trait                   description                url  
#>   <chr>     <chr>       <chr>                   <chr>                      <chr>
#> 1 PGS000010 EFO_0001645 coronary artery disease Narrowing of the coronary… http…
#> 
#> Slot "stages_tally":
#> # A tibble: 2 × 4
#>   pgs_id    stage sample_size n_sample_sets
#>   <chr>     <chr>       <int>         <int>
#> 1 PGS000010 gwas        86995            NA
#> 2 PGS000010 eval           NA             6
#> 
#> Slot "ancestry_frequencies":
#> # A tibble: 2 × 4
#>   pgs_id    stage ancestry_class_symbol frequency
#>   <chr>     <chr> <chr>                     <dbl>
#> 1 PGS000010 gwas  EUR                         100
#> 2 PGS000010 eval  EUR                         100
#> 
#> Slot "multi_ancestry_composition":
#> # A tibble: 0 × 4
#> # … with 4 variables: pgs_id <chr>, stage <chr>,
#> #   multi_ancestry_class_symbol <chr>, ancestry_class_symbol <chr>
#> # ℹ Use `colnames()` to see all variable names
#>