Get PGS Catalog Traits

Retrieves traits via the PGS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all traits that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the traits that match simultaneously all criteria provided, then set set_operation to 'intersection'.

get_traits(
  efo_id = NULL,
  trait_term = NULL,
  exact_term = TRUE,
  include_children = FALSE,
  set_operation = "union",
  interactive = TRUE,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

efo_id

A character vector of EFO identifiers.

trait_term

A character vector of terms to be matched against trait identifiers (efo_id), trait descriptions, synonyms thereof, externally mapped terms, or even trait categories.

exact_term

A logical value, indicating whether to match the trait_term exactly (TRUE) or not (FALSE).

include_children

A logical value, indicating whether to include child traits or not.

set_operation

Either 'union' or 'intersection'. This tells how performance metrics retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same sample sets found with different criteria.

interactive

A logical. If all traits are requested, whether to ask interactively if we really want to proceed.

verbose

A logical indicating whether the function should be verbose about the different queries or not.

warnings

A logical indicating whether to print warnings, if any.

progress_bar

Whether to show a progress bar indicating download progress from the REST API server.

Value

A traits object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search.

Examples

# Get a trait by its EFO identifier
get_traits(efo_id = 'EFO_0004631')
#> An object of class "traits"
#> Slot "traits":
#> # A tibble: 1 × 6
#>   efo_id      parent_efo_id is_child trait                 description     url  
#>   <chr>       <chr>         <lgl>    <chr>                 <chr>           <chr>
#> 1 EFO_0004631 NA            FALSE    vitamin D measurement "A quantificat… http…
#> 
#> Slot "pgs_ids":
#> # A tibble: 12 × 4
#>    efo_id      parent_efo_id is_child pgs_id   
#>    <chr>       <chr>         <lgl>    <chr>    
#>  1 EFO_0004631 NA            FALSE    PGS000702
#>  2 EFO_0004631 NA            FALSE    PGS000882
#>  3 EFO_0004631 NA            FALSE    PGS001982
#>  4 EFO_0004631 NA            FALSE    PGS002200
#>  5 EFO_0004631 NA            FALSE    PGS002355
#>  6 EFO_0004631 NA            FALSE    PGS002427
#>  7 EFO_0004631 NA            FALSE    PGS002476
#>  8 EFO_0004631 NA            FALSE    PGS002525
#>  9 EFO_0004631 NA            FALSE    PGS002574
#> 10 EFO_0004631 NA            FALSE    PGS002623
#> 11 EFO_0004631 NA            FALSE    PGS002672
#> 12 EFO_0004631 NA            FALSE    PGS002721
#> 
#> Slot "child_pgs_ids":
#> # A tibble: 0 × 4
#> # … with 4 variables: efo_id <chr>, parent_efo_id <chr>, is_child <lgl>,
#> #   child_pgs_id <chr>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "trait_categories":
#> # A tibble: 1 × 4
#>   efo_id      parent_efo_id is_child trait_categories 
#>   <chr>       <chr>         <lgl>    <chr>            
#> 1 EFO_0004631 NA            FALSE    Other measurement
#> 
#> Slot "trait_synonyms":
#> # A tibble: 0 × 4
#> # … with 4 variables: efo_id <chr>, parent_efo_id <chr>, is_child <lgl>,
#> #   trait_synonyms <chr>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "trait_mapped_terms":
#> # A tibble: 2 × 4
#>   efo_id      parent_efo_id is_child trait_mapped_terms
#>   <chr>       <chr>         <lgl>    <chr>             
#> 1 EFO_0004631 NA            FALSE    NCIt:C84818       
#> 2 EFO_0004631 NA            FALSE    SNOMEDCT:83729008 
#> 

# Get a trait by matching a term in EFO identifier (`efo_id`), label,
# description synonyms, categories, or external mapped terms
get_traits(trait_term = 'stroke', exact_term = FALSE)
#> An object of class "traits"
#> Slot "traits":
#> # A tibble: 3 × 6
#>   efo_id      parent_efo_id is_child trait                         descr…¹ url  
#>   <chr>       <chr>         <lgl>    <chr>                         <chr>   <chr>
#> 1 HP_0002140  NA            FALSE    Ischemic stroke               Acute … http…
#> 2 EFO_0010555 NA            FALSE    left ventricular stroke volu… Quanti… http…
#> 3 EFO_0000712 NA            FALSE    stroke                        A sudd… http…
#> # … with abbreviated variable name ¹​description
#> 
#> Slot "pgs_ids":
#> # A tibble: 9 × 4
#>   efo_id      parent_efo_id is_child pgs_id   
#>   <chr>       <chr>         <lgl>    <chr>    
#> 1 HP_0002140  NA            FALSE    PGS000039
#> 2 HP_0002140  NA            FALSE    PGS000665
#> 3 HP_0002140  NA            FALSE    PGS000911
#> 4 EFO_0010555 NA            FALSE    PGS001413
#> 5 EFO_0000712 NA            FALSE    PGS000038
#> 6 EFO_0000712 NA            FALSE    PGS000039
#> 7 EFO_0000712 NA            FALSE    PGS000665
#> 8 EFO_0000712 NA            FALSE    PGS000911
#> 9 EFO_0000712 NA            FALSE    PGS002259
#> 
#> Slot "child_pgs_ids":
#> # A tibble: 0 × 4
#> # … with 4 variables: efo_id <chr>, parent_efo_id <chr>, is_child <lgl>,
#> #   child_pgs_id <chr>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "trait_categories":
#> # A tibble: 4 × 4
#>   efo_id      parent_efo_id is_child trait_categories          
#>   <chr>       <chr>         <lgl>    <chr>                     
#> 1 HP_0002140  NA            FALSE    Other trait               
#> 2 EFO_0010555 NA            FALSE    Cardiovascular measurement
#> 3 EFO_0000712 NA            FALSE    Cardiovascular disease    
#> 4 EFO_0000712 NA            FALSE    Neurological disorder     
#> 
#> Slot "trait_synonyms":
#> # A tibble: 54 × 4
#>    efo_id      parent_efo_id is_child trait_synonyms                 
#>    <chr>       <chr>         <lgl>    <chr>                          
#>  1 HP_0002140  NA            FALSE    Ischaemic stroke               
#>  2 EFO_0000712 NA            FALSE    Acute Cerebrovascular Accident 
#>  3 EFO_0000712 NA            FALSE    Acute Cerebrovascular Accidents
#>  4 EFO_0000712 NA            FALSE    Acute Stroke                   
#>  5 EFO_0000712 NA            FALSE    Acute Strokes                  
#>  6 EFO_0000712 NA            FALSE    Apoplexy                       
#>  7 EFO_0000712 NA            FALSE    Apoplexy, Cerebrovascular      
#>  8 EFO_0000712 NA            FALSE    Brain Vascular Accident        
#>  9 EFO_0000712 NA            FALSE    Brain Vascular Accidents       
#> 10 EFO_0000712 NA            FALSE    CEREBROVASCULAR ACCIDENT, (CVA)
#> # … with 44 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "trait_mapped_terms":
#> # A tibble: 16 × 4
#>    efo_id      parent_efo_id is_child trait_mapped_terms   
#>    <chr>       <chr>         <lgl>    <chr>                
#>  1 HP_0002140  NA            FALSE    SNOMEDCT_US:422504002
#>  2 HP_0002140  NA            FALSE    UMLS:C0948008        
#>  3 EFO_0010555 NA            FALSE    PMID:31554410        
#>  4 EFO_0000712 NA            FALSE    HP:0001297           
#>  5 EFO_0000712 NA            FALSE    ICD10:I64            
#>  6 EFO_0000712 NA            FALSE    MESH:D020521         
#>  7 EFO_0000712 NA            FALSE    MONDO:0005098        
#>  8 EFO_0000712 NA            FALSE    MeSH:D020521         
#>  9 EFO_0000712 NA            FALSE    MedDRA:10042244      
#> 10 EFO_0000712 NA            FALSE    NCIT:C3390           
#> 11 EFO_0000712 NA            FALSE    NCIt:C3390           
#> 12 EFO_0000712 NA            FALSE    NIFSTD:birnlex_12783 
#> 13 EFO_0000712 NA            FALSE    OMIM:601367          
#> 14 EFO_0000712 NA            FALSE    SCTID:230690007      
#> 15 EFO_0000712 NA            FALSE    SNOMEDCT:230690007   
#> 16 EFO_0000712 NA            FALSE    SNOMEDCT:422504002   
#> 

# Get a trait matching its name (`trait`) exactly (default)
get_traits(trait_term = 'stroke', exact_term = TRUE)
#> An object of class "traits"
#> Slot "traits":
#> # A tibble: 1 × 6
#>   efo_id      parent_efo_id is_child trait  description                    url  
#>   <chr>       <chr>         <lgl>    <chr>  <chr>                          <chr>
#> 1 EFO_0000712 NA            FALSE    stroke A sudden loss of neurological… http…
#> 
#> Slot "pgs_ids":
#> # A tibble: 5 × 4
#>   efo_id      parent_efo_id is_child pgs_id   
#>   <chr>       <chr>         <lgl>    <chr>    
#> 1 EFO_0000712 NA            FALSE    PGS000038
#> 2 EFO_0000712 NA            FALSE    PGS000039
#> 3 EFO_0000712 NA            FALSE    PGS000665
#> 4 EFO_0000712 NA            FALSE    PGS000911
#> 5 EFO_0000712 NA            FALSE    PGS002259
#> 
#> Slot "child_pgs_ids":
#> # A tibble: 0 × 4
#> # … with 4 variables: efo_id <chr>, parent_efo_id <chr>, is_child <lgl>,
#> #   child_pgs_id <chr>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "trait_categories":
#> # A tibble: 2 × 4
#>   efo_id      parent_efo_id is_child trait_categories      
#>   <chr>       <chr>         <lgl>    <chr>                 
#> 1 EFO_0000712 NA            FALSE    Cardiovascular disease
#> 2 EFO_0000712 NA            FALSE    Neurological disorder 
#> 
#> Slot "trait_synonyms":
#> # A tibble: 53 × 4
#>    efo_id      parent_efo_id is_child trait_synonyms                 
#>    <chr>       <chr>         <lgl>    <chr>                          
#>  1 EFO_0000712 NA            FALSE    Acute Cerebrovascular Accident 
#>  2 EFO_0000712 NA            FALSE    Acute Cerebrovascular Accidents
#>  3 EFO_0000712 NA            FALSE    Acute Stroke                   
#>  4 EFO_0000712 NA            FALSE    Acute Strokes                  
#>  5 EFO_0000712 NA            FALSE    Apoplexy                       
#>  6 EFO_0000712 NA            FALSE    Apoplexy, Cerebrovascular      
#>  7 EFO_0000712 NA            FALSE    Brain Vascular Accident        
#>  8 EFO_0000712 NA            FALSE    Brain Vascular Accidents       
#>  9 EFO_0000712 NA            FALSE    CEREBROVASCULAR ACCIDENT, (CVA)
#> 10 EFO_0000712 NA            FALSE    CVA                            
#> # … with 43 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "trait_mapped_terms":
#> # A tibble: 13 × 4
#>    efo_id      parent_efo_id is_child trait_mapped_terms  
#>    <chr>       <chr>         <lgl>    <chr>               
#>  1 EFO_0000712 NA            FALSE    HP:0001297          
#>  2 EFO_0000712 NA            FALSE    ICD10:I64           
#>  3 EFO_0000712 NA            FALSE    MESH:D020521        
#>  4 EFO_0000712 NA            FALSE    MONDO:0005098       
#>  5 EFO_0000712 NA            FALSE    MeSH:D020521        
#>  6 EFO_0000712 NA            FALSE    MedDRA:10042244     
#>  7 EFO_0000712 NA            FALSE    NCIT:C3390          
#>  8 EFO_0000712 NA            FALSE    NCIt:C3390          
#>  9 EFO_0000712 NA            FALSE    NIFSTD:birnlex_12783
#> 10 EFO_0000712 NA            FALSE    OMIM:601367         
#> 11 EFO_0000712 NA            FALSE    SCTID:230690007     
#> 12 EFO_0000712 NA            FALSE    SNOMEDCT:230690007  
#> 13 EFO_0000712 NA            FALSE    SNOMEDCT:422504002  
#> 

# Get traits, excluding its children traits (default)
get_traits(trait_term = 'breast cancer')
#> An object of class "traits"
#> Slot "traits":
#> # A tibble: 1 × 6
#>   efo_id        parent_efo_id is_child trait         description           url  
#>   <chr>         <chr>         <lgl>    <chr>         <chr>                 <chr>
#> 1 MONDO_0007254 NA            FALSE    breast cancer A primary or metasta… http…
#> 
#> Slot "pgs_ids":
#> # A tibble: 0 × 4
#> # … with 4 variables: efo_id <chr>, parent_efo_id <chr>, is_child <lgl>,
#> #   pgs_id <chr>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "child_pgs_ids":
#> # A tibble: 112 × 4
#>    efo_id        parent_efo_id is_child child_pgs_id
#>    <chr>         <chr>         <lgl>    <chr>       
#>  1 MONDO_0007254 NA            FALSE    PGS000001   
#>  2 MONDO_0007254 NA            FALSE    PGS000002   
#>  3 MONDO_0007254 NA            FALSE    PGS000003   
#>  4 MONDO_0007254 NA            FALSE    PGS000004   
#>  5 MONDO_0007254 NA            FALSE    PGS000005   
#>  6 MONDO_0007254 NA            FALSE    PGS000006   
#>  7 MONDO_0007254 NA            FALSE    PGS000007   
#>  8 MONDO_0007254 NA            FALSE    PGS000008   
#>  9 MONDO_0007254 NA            FALSE    PGS000009   
#> 10 MONDO_0007254 NA            FALSE    PGS000015   
#> # … with 102 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "trait_categories":
#> # A tibble: 1 × 4
#>   efo_id        parent_efo_id is_child trait_categories
#>   <chr>         <chr>         <lgl>    <chr>           
#> 1 MONDO_0007254 NA            FALSE    Cancer          
#> 
#> Slot "trait_synonyms":
#> # A tibble: 10 × 4
#>    efo_id        parent_efo_id is_child trait_synonyms                  
#>    <chr>         <chr>         <lgl>    <chr>                           
#>  1 MONDO_0007254 NA            FALSE    BC                              
#>  2 MONDO_0007254 NA            FALSE    breast cancer                   
#>  3 MONDO_0007254 NA            FALSE    cancer of breast                
#>  4 MONDO_0007254 NA            FALSE    malignant breast neoplasm       
#>  5 MONDO_0007254 NA            FALSE    malignant breast tumor          
#>  6 MONDO_0007254 NA            FALSE    malignant neoplasm of breast    
#>  7 MONDO_0007254 NA            FALSE    malignant neoplasm of the breast
#>  8 MONDO_0007254 NA            FALSE    malignant tumor of breast       
#>  9 MONDO_0007254 NA            FALSE    malignant tumor of the breast   
#> 10 MONDO_0007254 NA            FALSE    mammary cancer                  
#> 
#> Slot "trait_mapped_terms":
#> # A tibble: 5 × 4
#>   efo_id        parent_efo_id is_child trait_mapped_terms
#>   <chr>         <chr>         <lgl>    <chr>             
#> 1 MONDO_0007254 NA            FALSE    DOID:1612         
#> 2 MONDO_0007254 NA            FALSE    ICD10CM:C50       
#> 3 MONDO_0007254 NA            FALSE    ICD9:174.8        
#> 4 MONDO_0007254 NA            FALSE    NCIT:C9335        
#> 5 MONDO_0007254 NA            FALSE    SCTID:254837009   
#> 

# Get traits, including its children traits (check column `is_child` for
# child traits)
get_traits(trait_term = 'breast cancer', include_children = TRUE)
#> An object of class "traits"
#> Slot "traits":
#> # A tibble: 9 × 6
#>   efo_id        parent_efo_id is_child trait                       descr…¹ url  
#>   <chr>         <chr>         <lgl>    <chr>                       <chr>   <chr>
#> 1 MONDO_0007254 NA            FALSE    breast cancer               A prim… http…
#> 2 MONDO_0021115 MONDO_0007254 TRUE     luminal B breast carcinoma  A biol… http…
#> 3 MONDO_0021116 MONDO_0007254 TRUE     luminal A breast carcinoma  A biol… http…
#> 4 EFO_1000649   MONDO_0007254 TRUE     estrogen-receptor positive… A subt… http…
#> 5 EFO_1000650   MONDO_0007254 TRUE     estrogen-receptor negative… A subt… http…
#> 6 EFO_1000294   MONDO_0007254 TRUE     HER2 Positive Breast Carci… A biol… http…
#> 7 EFO_0005537   MONDO_0007254 TRUE     triple-negative breast can… An inv… http…
#> 8 EFO_0000305   MONDO_0007254 TRUE     breast carcinoma            A carc… http…
#> 9 EFO_0009780   MONDO_0007254 TRUE     HER2 negative breast carci… A biol… http…
#> # … with abbreviated variable name ¹​description
#> 
#> Slot "pgs_ids":
#> # A tibble: 112 × 4
#>    efo_id        parent_efo_id is_child pgs_id   
#>    <chr>         <chr>         <lgl>    <chr>    
#>  1 MONDO_0021115 MONDO_0007254 TRUE     PGS000214
#>  2 MONDO_0021116 MONDO_0007254 TRUE     PGS000212
#>  3 EFO_1000649   MONDO_0007254 TRUE     PGS000002
#>  4 EFO_1000649   MONDO_0007254 TRUE     PGS000005
#>  5 EFO_1000649   MONDO_0007254 TRUE     PGS000008
#>  6 EFO_1000649   MONDO_0007254 TRUE     PGS000046
#>  7 EFO_1000649   MONDO_0007254 TRUE     PGS000347
#>  8 EFO_1000649   MONDO_0007254 TRUE     PGS000774
#>  9 EFO_1000650   MONDO_0007254 TRUE     PGS000003
#> 10 EFO_1000650   MONDO_0007254 TRUE     PGS000006
#> # … with 102 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "child_pgs_ids":
#> # A tibble: 131 × 4
#>    efo_id        parent_efo_id is_child child_pgs_id
#>    <chr>         <chr>         <lgl>    <chr>       
#>  1 MONDO_0007254 NA            FALSE    PGS000001   
#>  2 MONDO_0007254 NA            FALSE    PGS000002   
#>  3 MONDO_0007254 NA            FALSE    PGS000003   
#>  4 MONDO_0007254 NA            FALSE    PGS000004   
#>  5 MONDO_0007254 NA            FALSE    PGS000005   
#>  6 MONDO_0007254 NA            FALSE    PGS000006   
#>  7 MONDO_0007254 NA            FALSE    PGS000007   
#>  8 MONDO_0007254 NA            FALSE    PGS000008   
#>  9 MONDO_0007254 NA            FALSE    PGS000009   
#> 10 MONDO_0007254 NA            FALSE    PGS000015   
#> # … with 121 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "trait_categories":
#> # A tibble: 9 × 4
#>   efo_id        parent_efo_id is_child trait_categories
#>   <chr>         <chr>         <lgl>    <chr>           
#> 1 MONDO_0007254 NA            FALSE    Cancer          
#> 2 MONDO_0021115 MONDO_0007254 TRUE     Cancer          
#> 3 MONDO_0021116 MONDO_0007254 TRUE     Cancer          
#> 4 EFO_1000649   MONDO_0007254 TRUE     Cancer          
#> 5 EFO_1000650   MONDO_0007254 TRUE     Cancer          
#> 6 EFO_1000294   MONDO_0007254 TRUE     Cancer          
#> 7 EFO_0005537   MONDO_0007254 TRUE     Cancer          
#> 8 EFO_0000305   MONDO_0007254 TRUE     Cancer          
#> 9 EFO_0009780   MONDO_0007254 TRUE     Cancer          
#> 
#> Slot "trait_synonyms":
#> # A tibble: 40 × 4
#>    efo_id        parent_efo_id is_child trait_synonyms                  
#>    <chr>         <chr>         <lgl>    <chr>                           
#>  1 MONDO_0007254 NA            FALSE    BC                              
#>  2 MONDO_0007254 NA            FALSE    breast cancer                   
#>  3 MONDO_0007254 NA            FALSE    cancer of breast                
#>  4 MONDO_0007254 NA            FALSE    malignant breast neoplasm       
#>  5 MONDO_0007254 NA            FALSE    malignant breast tumor          
#>  6 MONDO_0007254 NA            FALSE    malignant neoplasm of breast    
#>  7 MONDO_0007254 NA            FALSE    malignant neoplasm of the breast
#>  8 MONDO_0007254 NA            FALSE    malignant tumor of breast       
#>  9 MONDO_0007254 NA            FALSE    malignant tumor of the breast   
#> 10 MONDO_0007254 NA            FALSE    mammary cancer                  
#> # … with 30 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "trait_mapped_terms":
#> # A tibble: 38 × 4
#>    efo_id        parent_efo_id is_child trait_mapped_terms
#>    <chr>         <chr>         <lgl>    <chr>             
#>  1 MONDO_0007254 NA            FALSE    DOID:1612         
#>  2 MONDO_0007254 NA            FALSE    ICD10CM:C50       
#>  3 MONDO_0007254 NA            FALSE    ICD9:174.8        
#>  4 MONDO_0007254 NA            FALSE    NCIT:C9335        
#>  5 MONDO_0007254 NA            FALSE    SCTID:254837009   
#>  6 MONDO_0021115 MONDO_0007254 TRUE     NCIT:C53555       
#>  7 MONDO_0021115 MONDO_0007254 TRUE     UMLS:C3642346     
#>  8 MONDO_0021116 MONDO_0007254 TRUE     NCIT:C53554       
#>  9 MONDO_0021116 MONDO_0007254 TRUE     UMLS:C3642345     
#> 10 EFO_1000649   MONDO_0007254 TRUE     DOID:0060075      
#> # … with 28 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>