Study stages

A dataset containing the various study stages assigned to samples in the PGS Catalog.

stages

Format

A data frame with 5 stages (rows) and 4 columns:

stage

Study stage.

symbol

One-letter symbol for the stage, or a comma separated combination thereof.

name

Stage name.

definition

Stage description.

Source

https://www.pgscatalog.org/docs/ancestry

Examples

stages

#> # A tibble: 5 × 4
#>   stage     symbol name                 definition                              
#>   <chr>     <chr>  <chr>                <chr>                                   
#> 1 gwas      G      Source of Variant A… "Describes the samples used to define t…
#> 2 dev       D      Score Development/T… "Describes the samples used to develop …
#> 3 eval      E      PGS Evaluation       "Information about the samples used in …
#> 4 gwas/dev  G,D    Development          "Combination of the stages \"gwas\" and…
#> 5 gwas/dev… G,D,E  All stages           "Combination of all the stages: \"gwas\…