Ancestry categories and classes

A dataset containing the ancestry categories defined in NHGRI-EBI GWAS Catalog framework (Table 1, doi:10.1186/s13059-018-1396-2 ). Ancestry categories are assigned to samples with distinct and well-defined patterns of genetic variation. You will find these categories in the variable ancestry_category of the following objects: scores, performance_metrics and sample_sets. Ancestry categories (ancestry_category) are further clustered into ancestry classes (ancestry_class).

Usage

ancestry_categories

Format

A data frame with 19 ancestry categories (rows) and 6 columns:

ancestry_category

Ancestry category.

ancestry_class

To reduce the complexity associated with the many ancestry categories, some have been merged into higher-level groupings (ancestry_class). These groupings represent the current breadth of data in the PGS Catalog and are likely to change as more data is added.

ancestry_class_symbol

3-letter code for the ancestry_class e.g. "EUR" or "MAE".

ancestry_class_colour

Hexadecimal colour code associated with ancestry groupings (ancestry_class). This can be useful when visually communicating about ancestries.

definition

Description of the ancestry category.

examples

Examples of detailed descriptions of sample ancestries included in the category.

Source

Table 1 of Moralles et al. (2018):

doi:10.1186/s13059-018-1396-2

PGS Catalog Ancestry Documentation:

http://www.pgscatalog.org/docs/ancestry/

Examples

ancestry_categories
#> # A tibble: 19 × 6
#>    ancestry_category                     ances…¹ ances…² ances…³ defin…⁴ examp…⁵
#>    <chr>                                 <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 Aboriginal Australian                 Additi… OTH     #999999 "Inclu… Martu …
#>  2 African American or Afro-Caribbean    African AFR     #FFD900 "Inclu… Africa…
#>  3 African unspecified                   African AFR     #FFD900 "Inclu… Africa…
#>  4 Asian unspecified                     Additi… ASN     #B15928 "Inclu… Asian,…
#>  5 Central Asian                         Additi… ASN     #B15928 "Inclu… Silk R…
#>  6 East Asian                            East A… EAS     #4DAF4A "Inclu… Chines…
#>  7 European                              Europe… EUR     #377EB8 "Inclu… Spanis…
#>  8 Greater Middle Eastern (Middle Easte… Greate… GME     #00CED1 "Inclu… Tunisi…
#>  9 Hispanic or Latin American            Hispan… AMR     #E41A1C "Inclu… Brazil…
#> 10 Native American                       Additi… OTH     #999999 "Inclu… Pima I…
#> 11 Not reported                          Ancest… NR      #BBBBBB "Inclu… NA     
#> 12 Oceanian                              Additi… OTH     #999999 "Inclu… Solomo…
#> 13 Other                                 Additi… OTH     #999999 "Inclu… Surina…
#> 14 Other admixed ancestry                Additi… OTH     #999999 "Inclu… NA     
#> 15 South Asian                           South … SAS     #984EA3 "Inclu… Bangla…
#> 16 South East Asian                      Additi… ASN     #B15928 "Inclu… Thai, …
#> 17 Sub-Saharan African                   African AFR     #FFD900 "Inclu… Yoruba…
#> 18 Multi-Ancestry (including Europeans)  Multi-… MAE     #A6CEE3 "Combi… NA     
#> 19 Multi-Ancestry (excluding Europeans)  Multi-… MAO     #FF7F00 "Combi… NA     
#> # … with abbreviated variable names ¹​ancestry_class, ²​ancestry_class_symbol,
#> #   ³​ancestry_class_colour, ⁴​definition, ⁵​examples