Ancestry categories

A dataset containing the ancestry categories defined in NHGRI-EBI GWAS Catalog framework (Table 1, Morales et al. (2018)). Ancestry categories are assigned to samples with distinct and well-defined patterns of genetic variation. You will find these categories in the variable ancestry of the following objects: scores, performance_metrics and sample_sets.

ancestry_categories

Format

A data frame with 17 rows and 3 columns:

ancestry

Ancestry category.

definition

Description of the ancestry category.

examples

Examples of detailed descriptions for samples included in the category.

Source

https://genomebiology.biomedcentral.com/articles/10.1186/s13059-018-1396-2/tables/1

Examples

ancestry_categories

#> # A tibble: 17 x 3
#>    ancestry                    definition                       examples        
#>    <chr>                       <chr>                            <chr>           
#>  1 Aboriginal Australian       "Includes individuals who eithe… Martu Australia…
#>  2 African American or Afro-C… "Includes individuals who eithe… African America…
#>  3 African unspecified         "Includes individuals that eith… African, non-Hi…
#>  4 Asian unspecified           "Includes individuals that eith… Asian, Asian Am…
#>  5 Central Asian               "Includes individuals who eithe… Silk Road (foun…
#>  6 East Asian                  "Includes individuals who eithe… Chinese, Japane…
#>  7 European                    "Includes individuals who eithe… Spanish, Swedish
#>  8 Greater Middle Eastern (Mi… "Includes individuals who self-… Tunisian, Arab,…
#>  9 Hispanic or Latin American  "Includes individuals who eithe… Brazilian, Mexi…
#> 10 Native American             "Includes indigenous individual… Pima Indian, Pl…
#> 11 Not reported                "Includes individuals for which… NA              
#> 12 Oceanian                    "Includes individuals that eith… Solomon Islande…
#> 13 Other                       "Includes individuals where an … Surinamese, Rus…
#> 14 Other admixed ancestry      "Includes individuals who eithe… NA              
#> 15 South Asian                 "Includes individuals who eithe… Bangladeshi, Sr…
#> 16 South East Asian            "Includes individuals who eithe… Thai, Malay     
#> 17 Sub-Saharan African         "Includes individuals who eithe… Yoruban, Gambian