Ancestry categories and classes

A dataset containing the ancestry categories defined in NHGRI-EBI GWAS Catalog framework (Table 1, Morales et al. (2018)). Ancestry categories are assigned to samples with distinct and well-defined patterns of genetic variation. You will find these categories in the variable ancestry_category of the following objects: scores, performance_metrics and sample_sets. Ancestry categories (ancestry_category) are further clustered into ancestry classes (ancestry_class).

ancestry_categories

Format

A data frame with 19 ancestry categories (rows) and 6 columns:

ancestry_category

Ancestry category.

ancestry_class

To reduce the complexity associated with the many ancestry categories, some have been merged into higher-level groupings (ancestry_class). These groupings represent the current breadth of data in the PGS Catalog and are likely to change as more data is added.

ancestry_class_symbol

3-letter code for the ancestry_class e.g. "EUR" or "MAE".

ancestry_class_colour

Hexadecimal colour code associated with ancestry groupings (ancestry_class). This can be useful when visually communicating about ancestries.

definition

Description of the ancestry category.

examples

Examples of detailed descriptions of sample ancestries included in the category.

Source

Table 1 of Morales et al. (2018):

https://genomebiology.biomedcentral.com/articles/10.1186/s13059-018-1396-2/tables/1

PGS Catalog Ancestry Documentation:

http://www.pgscatalog.org/docs/ancestry/

Examples

ancestry_categories

#> # A tibble: 19 × 6
#>    ancestry_category ancestry_class ancestry_class_… ancestry_class_… definition
#>    <chr>             <chr>          <chr>            <chr>            <chr>     
#>  1 Aboriginal Austr… Additional Di… OTH              #999999          "Includes…
#>  2 African American… African        AFR              #FFD900          "Includes…
#>  3 African unspecif… African        AFR              #FFD900          "Includes…
#>  4 Asian unspecified Additional As… ASN              #B15928          "Includes…
#>  5 Central Asian     Additional As… ASN              #B15928          "Includes…
#>  6 East Asian        East Asian     EAS              #4DAF4A          "Includes…
#>  7 European          European       EUR              #377EB8          "Includes…
#>  8 Greater Middle E… Greater Middl… GME              #00CED1          "Includes…
#>  9 Hispanic or Lati… Hispanic or L… AMR              #E41A1C          "Includes…
#> 10 Native American   Additional Di… OTH              #999999          "Includes…
#> 11 Not reported      Ancestry Not … NR               #BBBBBB          "Includes…
#> 12 Oceanian          Additional Di… OTH              #999999          "Includes…
#> 13 Other             Additional Di… OTH              #999999          "Includes…
#> 14 Other admixed an… Additional Di… OTH              #999999          "Includes…
#> 15 South Asian       South Asian    SAS              #984EA3          "Includes…
#> 16 South East Asian  Additional As… ASN              #B15928          "Includes…
#> 17 Sub-Saharan Afri… African        AFR              #FFD900          "Includes…
#> 18 Multi-Ancestry (… Multi-Ancestr… MAE              #A6CEE3          "Combined…
#> 19 Multi-Ancestry (… Multi-Ancestr… MAO              #FF7F00          "Combined…
#> # … with 1 more variable: examples <chr>