Get PGS Catalog Performance Metrics

Retrieves performance metrics via the PGS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all performance metrics that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the associations that match simultaneously all criteria provided, then set set_operation to 'intersection'.

get_performance_metrics(
  ppm_id = NULL,
  pgs_id = NULL,
  set_operation = "union",
  interactive = TRUE,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

ppm_id

A character vector of PGS Catalog performance metrics accession identifiers.

pgs_id

A character vector of PGS Catalog score accession identifiers.

set_operation

Either 'union' or 'intersection'. This tells how performance metrics retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same performance metrics found with different criteria.

interactive

A logical. If all performance metrics are requested, whether to ask interactively if we really want to proceed.

verbose

A logical indicating whether the function should be verbose about the different queries or not.

warnings

A logical indicating whether to print warnings, if any.

progress_bar

Whether to show a progress bar as the queries are performed.

Value

A performance_metrics object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search.

Examples

# Get performance metrics catalogued with identifier 'PPM000001'
get_performance_metrics(ppm_id = 'PPM000001')
#> An object of class "performance_metrics"
#> Slot "performance_metrics":
#> # A tibble: 1 × 5
#>   ppm_id    pgs_id    reported_trait    covariates comments
#>   <chr>     <chr>     <chr>             <chr>      <chr>   
#> 1 PPM000001 PGS000001 All breast cancer NA         NA      
#> 
#> Slot "publications":
#> # A tibble: 1 × 8
#>   ppm_id    pgp_id    pubmed_id publication_date publication title autho…¹ doi  
#>   <chr>     <chr>     <chr>     <date>           <chr>       <chr> <chr>   <chr>
#> 1 PPM000001 PGP000001 25855707  2015-04-08       J Natl Can… Pred… Mavadd… 10.1…
#> # … with abbreviated variable name ¹​author_fullname
#> 
#> Slot "sample_sets":
#> # A tibble: 1 × 2
#>   ppm_id    pss_id   
#>   <chr>     <chr>    
#> 1 PPM000001 PSS000001
#> 
#> Slot "samples":
#> # A tibble: 1 × 16
#>   ppm_id    pss_id sampl…¹ stage sampl…² sampl…³ sampl…⁴ sampl…⁵ pheno…⁶ ances…⁷
#>   <chr>     <chr>    <int> <chr>   <int>   <int>   <int>   <dbl> <chr>   <chr>  
#> 1 PPM000001 PSS00…       1 eval    67054   33673   33381      NA All br… Europe…
#> # … with 6 more variables: ancestry <chr>, country <chr>,
#> #   ancestry_additional_description <chr>, study_id <chr>, pubmed_id <chr>,
#> #   cohorts_additional_description <chr>, and abbreviated variable names
#> #   ¹​sample_id, ²​sample_size, ³​sample_cases, ⁴​sample_controls,
#> #   ⁵​sample_percent_male, ⁶​phenotype_description, ⁷​ancestry_category
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "demographics":
#> # A tibble: 0 × 12
#> # … with 12 variables: ppm_id <chr>, pss_id <chr>, sample_id <int>,
#> #   variable <chr>, estimate_type <chr>, estimate <dbl>, unit <chr>,
#> #   variability_type <chr>, variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "cohorts":
#> # A tibble: 33 × 5
#>    ppm_id    pss_id    sample_id cohort_symbol cohort_name                      
#>    <chr>     <chr>         <int> <chr>         <chr>                            
#>  1 PPM000001 PSS000001         1 ABCFS         Australian Breast Cancer Family …
#>  2 PPM000001 PSS000001         1 MCCS          Melbourne Collaborative Cohort S…
#>  3 PPM000001 PSS000001         1 HMBCS         Hannover-Minsk Breast Cancer Stu…
#>  4 PPM000001 PSS000001         1 LMBC          Leuven Multidisciplinary Breast …
#>  5 PPM000001 PSS000001         1 MTLGEBCS      Montreal Gene-Environment Breast…
#>  6 PPM000001 PSS000001         1 CGPS          Copenhagen General Population St…
#>  7 PPM000001 PSS000001         1 KBCP          Kuopio Breast Cancer Project     
#>  8 PPM000001 PSS000001         1 OBCS          Oulu Breast Cancer Study         
#>  9 PPM000001 PSS000001         1 CECILE        CECILE Breast Cancer Study       
#> 10 PPM000001 PSS000001         1 BBCC          Bavarian Breast Cancer Cases and…
#> # … with 23 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "pgs_effect_sizes":
#> # A tibble: 1 × 11
#>   ppm_id   effec…¹ estim…² estim…³ estim…⁴ unit  varia…⁵ varia…⁶ inter…⁷ inter…⁸
#>   <chr>      <int> <chr>   <chr>     <dbl> <chr> <chr>     <dbl> <chr>     <dbl>
#> 1 PPM0000…       1 Odds R… OR         1.55 NA    se           NA ci         1.52
#> # … with 1 more variable: interval_upper <dbl>, and abbreviated variable names
#> #   ¹​effect_size_id, ²​estimate_type_long, ³​estimate_type, ⁴​estimate,
#> #   ⁵​variability_type, ⁶​variability, ⁷​interval_type, ⁸​interval_lower
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "pgs_classification_metrics":
#> # A tibble: 1 × 11
#>   ppm_id   class…¹ estim…² estim…³ estim…⁴ unit  varia…⁵ varia…⁶ inter…⁷ inter…⁸
#>   <chr>      <int> <chr>   <chr>     <dbl> <chr> <chr>     <dbl> <chr>     <dbl>
#> 1 PPM0000…       1 Concor… C-index   0.622 NA    se           NA ci        0.619
#> # … with 1 more variable: interval_upper <dbl>, and abbreviated variable names
#> #   ¹​classification_metrics_id, ²​estimate_type_long, ³​estimate_type, ⁴​estimate,
#> #   ⁵​variability_type, ⁶​variability, ⁷​interval_type, ⁸​interval_lower
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "pgs_other_metrics":
#> # A tibble: 0 × 11
#> # … with 11 variables: ppm_id <chr>, other_metrics_id <int>,
#> #   estimate_type_long <chr>, estimate_type <chr>, estimate <dbl>, unit <chr>,
#> #   variability_type <chr>, variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> # ℹ Use `colnames()` to see all variable names
#> 

# Get performance metrics associated with polygenic score id 'PGS000001'
get_performance_metrics(pgs_id = 'PGS000001')
#> An object of class "performance_metrics"
#> Slot "performance_metrics":
#> # A tibble: 13 × 5
#>    ppm_id    pgs_id    reported_trait                            covar…¹ comme…²
#>    <chr>     <chr>     <chr>                                     <chr>   <chr>  
#>  1 PPM000001 PGS000001 All breast cancer                         NA      NA     
#>  2 PPM000011 PGS000001 Invasive breast cancer                    study,… NA     
#>  3 PPM000114 PGS000001 Breast cancer in BRCA1 mutation carriers  Countr… NA     
#>  4 PPM000117 PGS000001 Breast cancer in BRCA2 mutation carriers  Countr… NA     
#>  5 PPM000944 PGS000001 Metachronous contralateral breast cancer  Country NA     
#>  6 PPM000945 PGS000001 Invasive metachronous contralateral brea… Country NA     
#>  7 PPM000961 PGS000001 Metachronous contralateral breast cancer  Country NA     
#>  8 PPM000962 PGS000001 Invasive metachronous contralateral brea… Country NA     
#>  9 PPM002150 PGS000001 Breast cancer in CHEK2 mutation carriers  Year o… Only 7…
#> 10 PPM002151 PGS000001 Breast cancer in CHEK2 mutation carriers  Year o… Only 7…
#> 11 PPM002152 PGS000001 Breast cancer in CHEK2 mutation carriers… Year o… Only 7…
#> 12 PPM002153 PGS000001 Breast cancer in CHEK2 mutation carriers… Year o… Only 7…
#> 13 PPM002154 PGS000001 Breast cancer in CHEK2 mutation carriers… Year o… Only 7…
#> # … with abbreviated variable names ¹​covariates, ²​comments
#> 
#> Slot "publications":
#> # A tibble: 13 × 8
#>    ppm_id    pgp_id    pubmed_id publication_date publicat…¹ title autho…² doi  
#>    <chr>     <chr>     <chr>     <date>           <chr>      <chr> <chr>   <chr>
#>  1 PPM000001 PGP000001 25855707  2015-04-08       J Natl Ca… Pred… Mavadd… 10.1…
#>  2 PPM000011 PGP000002 30554720  2018-12-13       Am J Hum … Poly… Mavadd… 10.1…
#>  3 PPM000114 PGP000033 28376175  2017-07-01       J Natl Ca… Eval… Kuchen… 10.1…
#>  4 PPM000117 PGP000033 28376175  2017-07-01       J Natl Ca… Eval… Kuchen… 10.1…
#>  5 PPM000944 PGP000109 33022221  2020-10-05       Am J Hum … Brea… Kramer… 10.1…
#>  6 PPM000945 PGP000109 33022221  2020-10-05       Am J Hum … Brea… Kramer… 10.1…
#>  7 PPM000961 PGP000109 33022221  2020-10-05       Am J Hum … Brea… Kramer… 10.1…
#>  8 PPM000962 PGP000109 33022221  2020-10-05       Am J Hum … Brea… Kramer… 10.1…
#>  9 PPM002150 PGP000198 33372680  2020-12-29       J Natl Ca… Perf… Borde J 10.1…
#> 10 PPM002151 PGP000198 33372680  2020-12-29       J Natl Ca… Perf… Borde J 10.1…
#> 11 PPM002152 PGP000198 33372680  2020-12-29       J Natl Ca… Perf… Borde J 10.1…
#> 12 PPM002153 PGP000198 33372680  2020-12-29       J Natl Ca… Perf… Borde J 10.1…
#> 13 PPM002154 PGP000198 33372680  2020-12-29       J Natl Ca… Perf… Borde J 10.1…
#> # … with abbreviated variable names ¹​publication, ²​author_fullname
#> 
#> Slot "sample_sets":
#> # A tibble: 13 × 2
#>    ppm_id    pss_id   
#>    <chr>     <chr>    
#>  1 PPM000001 PSS000001
#>  2 PPM000011 PSS000004
#>  3 PPM000114 PSS000070
#>  4 PPM000117 PSS000071
#>  5 PPM000944 PSS000484
#>  6 PPM000945 PSS000486
#>  7 PPM000961 PSS000484
#>  8 PPM000962 PSS000486
#>  9 PPM002150 PSS001054
#> 10 PPM002151 PSS001054
#> 11 PPM002152 PSS001054
#> 12 PPM002153 PSS001054
#> 13 PPM002154 PSS001054
#> 
#> Slot "samples":
#> # A tibble: 13 × 16
#>    ppm_id   pss_id sampl…¹ stage sampl…² sampl…³ sampl…⁴ sampl…⁵ pheno…⁶ ances…⁷
#>    <chr>    <chr>    <int> <chr>   <int>   <int>   <int>   <dbl> <chr>   <chr>  
#>  1 PPM0000… PSS00…       1 eval    67054   33673   33381      NA "All b… Europe…
#>  2 PPM0000… PSS00…       1 eval    29751   11428   18323       0 "Invas… Europe…
#>  3 PPM0001… PSS00…       1 eval    15252    7797    7455       0 "BRCA1… Europe…
#>  4 PPM0001… PSS00…       1 eval     8211    4330    3881       0 "BRCA2… Europe…
#>  5 PPM0009… PSS00…       1 eval    56068    1027   55041       0 "Women… Europe…
#>  6 PPM0009… PSS00…       1 eval    56068     923   55145       0 "Women… Europe…
#>  7 PPM0009… PSS00…       1 eval    56068    1027   55041       0 "Women… Europe…
#>  8 PPM0009… PSS00…       1 eval    56068     923   55145       0 "Women… Europe…
#>  9 PPM0021… PSS00…       1 eval      760     561     199       0 "All w… Europe…
#> 10 PPM0021… PSS00…       1 eval      760     561     199       0 "All w… Europe…
#> 11 PPM0021… PSS00…       1 eval      760     561     199       0 "All w… Europe…
#> 12 PPM0021… PSS00…       1 eval      760     561     199       0 "All w… Europe…
#> 13 PPM0021… PSS00…       1 eval      760     561     199       0 "All w… Europe…
#> # … with 6 more variables: ancestry <chr>, country <chr>,
#> #   ancestry_additional_description <chr>, study_id <chr>, pubmed_id <chr>,
#> #   cohorts_additional_description <chr>, and abbreviated variable names
#> #   ¹​sample_id, ²​sample_size, ³​sample_cases, ⁴​sample_controls,
#> #   ⁵​sample_percent_male, ⁶​phenotype_description, ⁷​ancestry_category
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "demographics":
#> # A tibble: 0 × 12
#> # … with 12 variables: ppm_id <chr>, pss_id <chr>, sample_id <int>,
#> #   variable <chr>, estimate_type <chr>, estimate <dbl>, unit <chr>,
#> #   variability_type <chr>, variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "cohorts":
#> # A tibble: 218 × 5
#>    ppm_id    pss_id    sample_id cohort_symbol cohort_name                      
#>    <chr>     <chr>         <int> <chr>         <chr>                            
#>  1 PPM000001 PSS000001         1 ABCFS         Australian Breast Cancer Family …
#>  2 PPM000001 PSS000001         1 MCCS          Melbourne Collaborative Cohort S…
#>  3 PPM000001 PSS000001         1 HMBCS         Hannover-Minsk Breast Cancer Stu…
#>  4 PPM000001 PSS000001         1 LMBC          Leuven Multidisciplinary Breast …
#>  5 PPM000001 PSS000001         1 MTLGEBCS      Montreal Gene-Environment Breast…
#>  6 PPM000001 PSS000001         1 CGPS          Copenhagen General Population St…
#>  7 PPM000001 PSS000001         1 KBCP          Kuopio Breast Cancer Project     
#>  8 PPM000001 PSS000001         1 OBCS          Oulu Breast Cancer Study         
#>  9 PPM000001 PSS000001         1 CECILE        CECILE Breast Cancer Study       
#> 10 PPM000001 PSS000001         1 BBCC          Bavarian Breast Cancer Cases and…
#> # … with 208 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> Slot "pgs_effect_sizes":
#> # A tibble: 13 × 11
#>    ppm_id  effec…¹ estim…² estim…³ estim…⁴ unit  varia…⁵ varia…⁶ inter…⁷ inter…⁸
#>    <chr>     <int> <chr>   <chr>     <dbl> <chr> <chr>     <dbl> <chr>     <dbl>
#>  1 PPM000…       1 Odds R… OR         1.55 NA    se           NA ci         1.52
#>  2 PPM000…       1 Odds R… OR         1.46 NA    se           NA ci         1.42
#>  3 PPM000…       1 Hazard… HR         1.13 NA    se           NA ci         1.1 
#>  4 PPM000…       1 Hazard… HR         1.22 NA    se           NA ci         1.17
#>  5 PPM000…       1 Hazard… HR         1.21 NA    se           NA ci         1.14
#>  6 PPM000…       1 Hazard… HR         1.21 NA    se           NA ci         1.13
#>  7 PPM000…       1 Hazard… HR         1.21 NA    se           NA ci         1.14
#>  8 PPM000…       1 Hazard… HR         1.21 NA    se           NA ci         1.13
#>  9 PPM002…       1 Hazard… HR         1.71 NA    se           NA ci         1.36
#> 10 PPM002…       1 Hazard… HR         2.29 NA    se           NA ci         1.56
#> 11 PPM002…       1 Hazard… HR         1.43 NA    se           NA ci         1.04
#> 12 PPM002…       1 Hazard… HR         2.32 NA    se           NA ci         1.69
#> 13 PPM002…       1 Hazard… HR         1.59 NA    se           NA ci         1.07
#> # … with 1 more variable: interval_upper <dbl>, and abbreviated variable names
#> #   ¹​effect_size_id, ²​estimate_type_long, ³​estimate_type, ⁴​estimate,
#> #   ⁵​variability_type, ⁶​variability, ⁷​interval_type, ⁸​interval_lower
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "pgs_classification_metrics":
#> # A tibble: 2 × 11
#>   ppm_id   class…¹ estim…² estim…³ estim…⁴ unit  varia…⁵ varia…⁶ inter…⁷ inter…⁸
#>   <chr>      <int> <chr>   <chr>     <dbl> <chr> <chr>     <dbl> <chr>     <dbl>
#> 1 PPM0000…       1 Concor… C-index   0.622 NA    se           NA ci        0.619
#> 2 PPM0000…       1 Area U… AUROC     0.603 NA    se           NA ci       NA    
#> # … with 1 more variable: interval_upper <dbl>, and abbreviated variable names
#> #   ¹​classification_metrics_id, ²​estimate_type_long, ³​estimate_type, ⁴​estimate,
#> #   ⁵​variability_type, ⁶​variability, ⁷​interval_type, ⁸​interval_lower
#> # ℹ Use `colnames()` to see all variable names
#> 
#> Slot "pgs_other_metrics":
#> # A tibble: 0 × 11
#> # … with 11 variables: ppm_id <chr>, other_metrics_id <int>,
#> #   estimate_type_long <chr>, estimate_type <chr>, estimate <dbl>, unit <chr>,
#> #   variability_type <chr>, variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> # ℹ Use `colnames()` to see all variable names
#> 

# To retrieve all catalogued performed metrics in PGS Catalog you simply
# leave the parameters `ppm_id` and `pgs_id` as `NULL`.
if (FALSE) {
get_performance_metrics()
}