Get PGS Catalog Performance Metrics

Retrieves performance metrics via the PGS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all performance metrics that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the associations that match simultaneously all criteria provided, then set set_operation to 'intersection'.

get_performance_metrics(
  ppm_id = NULL,
  pgs_id = NULL,
  set_operation = "union",
  interactive = TRUE,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

ppm_id
A character vector of PGS Catalog performance metrics accession identifiers.

pgs_id
A character vector of PGS Catalog score accession identifiers.

set_operation
Either 'union' or 'intersection'. This tells how performance metrics retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same performance metrics found with different criteria.

interactive
A logical. If all performance metrics are requested, whether to ask interactively if we really want to proceed.

verbose
A logical indicating whether the function should be verbose about the different queries or not.

warnings
A logical indicating whether to print warnings, if any.

progress_bar
Whether to show a progress bar as the queries are performed.

Value

A performance_metrics object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search.

Examples

# Get performance metrics catalogued with identifier 'PPM000001'
get_performance_metrics(ppm_id = 'PPM000001')

#> An object of class "performance_metrics"
#> Slot "performance_metrics":
#> # A tibble: 1 × 5
#>   ppm_id    pgs_id    reported_trait    covariates comments
#>   <chr>     <chr>     <chr>             <chr>      <chr>   
#> 1 PPM000001 PGS000001 All breast cancer NA         NA      
#> 
#> Slot "publications":
#> # A tibble: 1 × 8
#>   ppm_id    pgp_id    pubmed_id publication_date publication title author_fullname
#>   <chr>     <chr>     <chr>     <date>           <chr>       <chr> <chr>          
#> 1 PPM000001 PGP000001 25855707  2015-04-08       J Natl Can… Pred… Mavaddat N     
#> # … with 1 more variable: doi <chr>
#> 
#> Slot "sample_sets":
#> # A tibble: 1 × 2
#>   ppm_id    pss_id   
#>   <chr>     <chr>    
#> 1 PPM000001 PSS000001
#> 
#> Slot "samples":
#> # A tibble: 1 × 16
#>   ppm_id    pss_id    sample_id stage sample_size sample_cases sample_controls
#>   <chr>     <chr>         <int> <chr>       <int>        <int>           <int>
#> 1 PPM000001 PSS000001         1 eval        67054        33673           33381
#> # … with 9 more variables: sample_percent_male <dbl>,
#> #   phenotype_description <chr>, ancestry_category <chr>, ancestry <chr>,
#> #   country <chr>, ancestry_additional_description <chr>, study_id <chr>,
#> #   pubmed_id <chr>, cohorts_additional_description <chr>
#> 
#> Slot "demographics":
#> # A tibble: 0 × 12
#> # … with 12 variables: ppm_id <chr>, pss_id <chr>, sample_id <int>,
#> #   variable <chr>, estimate_type <chr>, estimate <dbl>, unit <chr>,
#> #   variability_type <chr>, variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> 
#> Slot "cohorts":
#> # A tibble: 33 × 5
#>    ppm_id    pss_id    sample_id cohort_symbol cohort_name                      
#>    <chr>     <chr>         <int> <chr>         <chr>                            
#>  1 PPM000001 PSS000001         1 DEMOKRITOS    Hellenic Cooperative Oncology Gr…
#>  2 PPM000001 PSS000001         1 SBCS          Shanghai Breast Cancer Study     
#>  3 PPM000001 PSS000001         1 SEARCH        Study of Epidemiology and Risk F…
#>  4 PPM000001 PSS000001         1 MCBCS         Mayo Clinic Breast Cancer Study  
#>  5 PPM000001 PSS000001         1 RPCI          Roswell Park Cancer Institute    
#>  6 PPM000001 PSS000001         1 OSU           The Stefanie Spielman Breast Ban…
#>  7 PPM000001 PSS000001         1 GENICA        Gene Environment Interaction and…
#>  8 PPM000001 PSS000001         1 MARIE         Mammary Carcinoma Risk Factor In…
#>  9 PPM000001 PSS000001         1 SKKDKFZS      St_dtisches Klinikum Karlsruhe D…
#> 10 PPM000001 PSS000001         1 BIGGS         Breast Cancer In Galway Genetic …
#> # … with 23 more rows
#> 
#> Slot "pgs_effect_sizes":
#> # A tibble: 1 × 11
#>   ppm_id    effect_size_id estimate_type_long estimate_type estimate unit 
#>   <chr>              <int> <chr>              <chr>            <dbl> <chr>
#> 1 PPM000001              1 Odds Ratio         OR                1.55 NA   
#> # … with 5 more variables: variability_type <chr>, variability <dbl>,
#> #   interval_type <chr>, interval_lower <dbl>, interval_upper <dbl>
#> 
#> Slot "pgs_classification_metrics":
#> # A tibble: 1 × 11
#>   ppm_id    classification_metr… estimate_type_long estimate_type estimate unit 
#>   <chr>                    <int> <chr>              <chr>            <dbl> <chr>
#> 1 PPM000001                    1 Concordance Stati… C-index          0.622 NA   
#> # … with 5 more variables: variability_type <chr>, variability <dbl>,
#> #   interval_type <chr>, interval_lower <dbl>, interval_upper <dbl>
#> 
#> Slot "pgs_other_metrics":
#> # A tibble: 0 × 11
#> # … with 11 variables: ppm_id <chr>, other_metrics_id <int>,
#> #   estimate_type_long <chr>, estimate_type <chr>, estimate <dbl>, unit <chr>,
#> #   variability_type <chr>, variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> 

# Get performance metrics associated with polygenic score id 'PGS000001'
get_performance_metrics(pgs_id = 'PGS000001')

#> An object of class "performance_metrics"
#> Slot "performance_metrics":
#> # A tibble: 13 × 5
#>    ppm_id    pgs_id    reported_trait        covariates         comments        
#>    <chr>     <chr>     <chr>                 <chr>              <chr>           
#>  1 PPM000001 PGS000001 All breast cancer     NA                 NA              
#>  2 PPM000011 PGS000001 Invasive breast canc… study, genetic PC… NA              
#>  3 PPM000114 PGS000001 Breast cancer in BRC… Country, birth ye… NA              
#>  4 PPM000117 PGS000001 Breast cancer in BRC… Country, birth ye… NA              
#>  5 PPM000944 PGS000001 Metachronous contral… Country            NA              
#>  6 PPM000945 PGS000001 Invasive metachronou… Country            NA              
#>  7 PPM000961 PGS000001 Metachronous contral… Country            NA              
#>  8 PPM000962 PGS000001 Invasive metachronou… Country            NA              
#>  9 PPM002150 PGS000001 Breast cancer in CHE… Year of birth, co… Only 70 of the …
#> 10 PPM002151 PGS000001 Breast cancer in CHE… Year of birth, co… Only 70 of the …
#> 11 PPM002152 PGS000001 Breast cancer in CHE… Year of birth, co… Only 70 of the …
#> 12 PPM002153 PGS000001 Breast cancer in CHE… Year of birth, co… Only 70 of the …
#> 13 PPM002154 PGS000001 Breast cancer in CHE… Year of birth, co… Only 70 of the …
#> 
#> Slot "publications":
#> # A tibble: 13 × 8
#>    ppm_id    pgp_id    pubmed_id publication_date publication title author_fullname
#>    <chr>     <chr>     <chr>     <date>           <chr>       <chr> <chr>          
#>  1 PPM000001 PGP000001 25855707  2015-04-08       J Natl Can… Pred… Mavaddat N     
#>  2 PPM000011 PGP000002 30554720  2018-12-13       Am J Hum G… Poly… Mavaddat N     
#>  3 PPM000114 PGP000033 28376175  2017-07-01       J Natl Can… Eval… Kuchenbaecker …
#>  4 PPM000117 PGP000033 28376175  2017-07-01       J Natl Can… Eval… Kuchenbaecker …
#>  5 PPM000944 PGP000109 33022221  2020-10-05       Am J Hum G… Brea… Kramer I       
#>  6 PPM000945 PGP000109 33022221  2020-10-05       Am J Hum G… Brea… Kramer I       
#>  7 PPM000961 PGP000109 33022221  2020-10-05       Am J Hum G… Brea… Kramer I       
#>  8 PPM000962 PGP000109 33022221  2020-10-05       Am J Hum G… Brea… Kramer I       
#>  9 PPM002150 PGP000198 33372680  2020-12-29       J Natl Can… Perf… Borde J        
#> 10 PPM002151 PGP000198 33372680  2020-12-29       J Natl Can… Perf… Borde J        
#> 11 PPM002152 PGP000198 33372680  2020-12-29       J Natl Can… Perf… Borde J        
#> 12 PPM002153 PGP000198 33372680  2020-12-29       J Natl Can… Perf… Borde J        
#> 13 PPM002154 PGP000198 33372680  2020-12-29       J Natl Can… Perf… Borde J        
#> # … with 1 more variable: doi <chr>
#> 
#> Slot "sample_sets":
#> # A tibble: 13 × 2
#>    ppm_id    pss_id   
#>    <chr>     <chr>    
#>  1 PPM000001 PSS000001
#>  2 PPM000011 PSS000004
#>  3 PPM000114 PSS000070
#>  4 PPM000117 PSS000071
#>  5 PPM000944 PSS000484
#>  6 PPM000945 PSS000486
#>  7 PPM000961 PSS000484
#>  8 PPM000962 PSS000486
#>  9 PPM002150 PSS001054
#> 10 PPM002151 PSS001054
#> 11 PPM002152 PSS001054
#> 12 PPM002153 PSS001054
#> 13 PPM002154 PSS001054
#> 
#> Slot "samples":
#> # A tibble: 13 × 16
#>    ppm_id    pss_id    sample_id stage sample_size sample_cases sample_controls
#>    <chr>     <chr>         <int> <chr>       <int>        <int>           <int>
#>  1 PPM000001 PSS000001         1 eval        67054        33673           33381
#>  2 PPM000011 PSS000004         1 eval        29751        11428           18323
#>  3 PPM000114 PSS000070         1 eval        15252         7797            7455
#>  4 PPM000117 PSS000071         1 eval         8211         4330            3881
#>  5 PPM000944 PSS000484         1 eval        56068         1027           55041
#>  6 PPM000945 PSS000486         1 eval        56068          923           55145
#>  7 PPM000961 PSS000484         1 eval        56068         1027           55041
#>  8 PPM000962 PSS000486         1 eval        56068          923           55145
#>  9 PPM002150 PSS001054         1 eval          760          561             199
#> 10 PPM002151 PSS001054         1 eval          760          561             199
#> 11 PPM002152 PSS001054         1 eval          760          561             199
#> 12 PPM002153 PSS001054         1 eval          760          561             199
#> 13 PPM002154 PSS001054         1 eval          760          561             199
#> # … with 9 more variables: sample_percent_male <dbl>,
#> #   phenotype_description <chr>, ancestry_category <chr>, ancestry <chr>,
#> #   country <chr>, ancestry_additional_description <chr>, study_id <chr>,
#> #   pubmed_id <chr>, cohorts_additional_description <chr>
#> 
#> Slot "demographics":
#> # A tibble: 0 × 12
#> # … with 12 variables: ppm_id <chr>, pss_id <chr>, sample_id <int>,
#> #   variable <chr>, estimate_type <chr>, estimate <dbl>, unit <chr>,
#> #   variability_type <chr>, variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> 
#> Slot "cohorts":
#> # A tibble: 218 × 5
#>    ppm_id    pss_id    sample_id cohort_symbol cohort_name                      
#>    <chr>     <chr>         <int> <chr>         <chr>                            
#>  1 PPM000001 PSS000001         1 ABCFS         Australian Breast Cancer Family …
#>  2 PPM000001 PSS000001         1 MCCS          Melbourne Collaborative Cohort S…
#>  3 PPM000001 PSS000001         1 HMBCS         Hannover-Minsk Breast Cancer Stu…
#>  4 PPM000001 PSS000001         1 LMBC          Leuven Multidisciplinary Breast …
#>  5 PPM000001 PSS000001         1 MTLGEBCS      Montreal Gene-Environment Breast…
#>  6 PPM000001 PSS000001         1 CGPS          Copenhagen General Population St…
#>  7 PPM000001 PSS000001         1 KBCP          Kuopio Breast Cancer Project     
#>  8 PPM000001 PSS000001         1 OBCS          Oulu Breast Cancer Study         
#>  9 PPM000001 PSS000001         1 CECILE        CECILE Breast Cancer Study       
#> 10 PPM000001 PSS000001         1 BBCC          Bavarian Breast Cancer Cases and…
#> # … with 208 more rows
#> 
#> Slot "pgs_effect_sizes":
#> # A tibble: 13 × 11
#>    ppm_id    effect_size_id estimate_type_long estimate_type estimate unit 
#>    <chr>              <int> <chr>              <chr>            <dbl> <chr>
#>  1 PPM000001              1 Odds Ratio         OR                1.55 NA   
#>  2 PPM000011              1 Odds Ratio         OR                1.46 NA   
#>  3 PPM000114              1 Hazard Ratio       HR                1.13 NA   
#>  4 PPM000117              1 Hazard Ratio       HR                1.22 NA   
#>  5 PPM000944              1 Hazard Ratio       HR                1.21 NA   
#>  6 PPM000945              1 Hazard Ratio       HR                1.21 NA   
#>  7 PPM000961              1 Hazard Ratio       HR                1.21 NA   
#>  8 PPM000962              1 Hazard Ratio       HR                1.21 NA   
#>  9 PPM002150              1 Hazard Ratio       HR                1.71 NA   
#> 10 PPM002151              1 Hazard Ratio       HR                2.29 NA   
#> 11 PPM002152              1 Hazard Ratio       HR                1.43 NA   
#> 12 PPM002153              1 Hazard Ratio       HR                2.32 NA   
#> 13 PPM002154              1 Hazard Ratio       HR                1.59 NA   
#> # … with 5 more variables: variability_type <chr>, variability <dbl>,
#> #   interval_type <chr>, interval_lower <dbl>, interval_upper <dbl>
#> 
#> Slot "pgs_classification_metrics":
#> # A tibble: 2 × 11
#>   ppm_id    classification_me… estimate_type_long   estimate_type estimate unit 
#>   <chr>                  <int> <chr>                <chr>            <dbl> <chr>
#> 1 PPM000001                  1 Concordance Statist… C-index          0.622 NA   
#> 2 PPM000011                  1 Area Under the Rece… AUROC            0.603 NA   
#> # … with 5 more variables: variability_type <chr>, variability <dbl>,
#> #   interval_type <chr>, interval_lower <dbl>, interval_upper <dbl>
#> 
#> Slot "pgs_other_metrics":
#> # A tibble: 0 × 11
#> # … with 11 variables: ppm_id <chr>, other_metrics_id <int>,
#> #   estimate_type_long <chr>, estimate_type <chr>, estimate <dbl>, unit <chr>,
#> #   variability_type <chr>, variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> 

# To retrieve all catalogued performed metrics in PGS Catalog you simply
# leave the parameters `ppm_id` and `pgs_id` as `NULL`.
if (FALSE) {
get_performance_metrics()
}