Get PGS Catalog Releases

This function retrieves PGS Catalog release information. Note that the columns pgs_id, ppm_id and pgp_id contain in each element a vector. These columns can be unnested using unnest_longer (see Examples).

get_releases(
  date = "latest",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

date
One or more dates formatted as "YYYY-MM-DD" for respective releases, "latest" for the latest release, or "all" for all releases.

verbose
Whether to print information about the underlying requests to the REST API server.

warnings
Whether to print warnings about the underlying requests to the REST API server.

progress_bar
Whether to show a progress bar indicating download progress from the REST API server.

Value

A data frame where each row is a release. Columns are:

date

Release date.

n_pgs

Number of released Polygenic Score (PGS) identifiers (pgs_id).

n_ppm

Number of released Performance Metric (PPM) identifiers (ppm_id).

n_pgp

Number of released PGS Catalog Publication (PGP) identifiers (pgp_id).

pgs_id

Released Polygenic Score (PGS) identifiers.

ppm_id

Released Performance Metric (PPM) identifiers.

pgp_id

Released PGS Catalog Publication (PGP) identifiers.

notes

News about the release.

Examples

# Get the latest release
get_releases()

#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 x 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2021-02-03    58   265     8 This release contains 58 new Score(s), 8 new Pub…
#> 
#> Slot "pgs_ids":
#> # A tibble: 58 x 2
#>    date       pgs_id   
#>    <date>     <chr>    
#>  1 2021-02-03 PGS000668
#>  2 2021-02-03 PGS000669
#>  3 2021-02-03 PGS000670
#>  4 2021-02-03 PGS000671
#>  5 2021-02-03 PGS000672
#>  6 2021-02-03 PGS000673
#>  7 2021-02-03 PGS000674
#>  8 2021-02-03 PGS000675
#>  9 2021-02-03 PGS000676
#> 10 2021-02-03 PGS000677
#> # … with 48 more rows
#> 
#> Slot "ppm_ids":
#> # A tibble: 265 x 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2021-02-03 PPM001396
#>  2 2021-02-03 PPM001397
#>  3 2021-02-03 PPM001398
#>  4 2021-02-03 PPM001399
#>  5 2021-02-03 PPM001400
#>  6 2021-02-03 PPM001401
#>  7 2021-02-03 PPM001402
#>  8 2021-02-03 PPM001403
#>  9 2021-02-03 PPM001404
#> 10 2021-02-03 PPM001405
#> # … with 255 more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 8 x 2
#>   date       pgp_id   
#>   <date>     <chr>    
#> 1 2021-02-03 PGP000128
#> 2 2021-02-03 PGP000129
#> 3 2021-02-03 PGP000130
#> 4 2021-02-03 PGP000132
#> 5 2021-02-03 PGP000133
#> 6 2021-02-03 PGP000134
#> 7 2021-02-03 PGP000135
#> 8 2021-02-03 PGP000136
#> 

get_releases(date = 'latest')

#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 x 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2021-02-03    58   265     8 This release contains 58 new Score(s), 8 new Pub…
#> 
#> Slot "pgs_ids":
#> # A tibble: 58 x 2
#>    date       pgs_id   
#>    <date>     <chr>    
#>  1 2021-02-03 PGS000668
#>  2 2021-02-03 PGS000669
#>  3 2021-02-03 PGS000670
#>  4 2021-02-03 PGS000671
#>  5 2021-02-03 PGS000672
#>  6 2021-02-03 PGS000673
#>  7 2021-02-03 PGS000674
#>  8 2021-02-03 PGS000675
#>  9 2021-02-03 PGS000676
#> 10 2021-02-03 PGS000677
#> # … with 48 more rows
#> 
#> Slot "ppm_ids":
#> # A tibble: 265 x 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2021-02-03 PPM001396
#>  2 2021-02-03 PPM001397
#>  3 2021-02-03 PPM001398
#>  4 2021-02-03 PPM001399
#>  5 2021-02-03 PPM001400
#>  6 2021-02-03 PPM001401
#>  7 2021-02-03 PPM001402
#>  8 2021-02-03 PPM001403
#>  9 2021-02-03 PPM001404
#> 10 2021-02-03 PPM001405
#> # … with 255 more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 8 x 2
#>   date       pgp_id   
#>   <date>     <chr>    
#> 1 2021-02-03 PGP000128
#> 2 2021-02-03 PGP000129
#> 3 2021-02-03 PGP000130
#> 4 2021-02-03 PGP000132
#> 5 2021-02-03 PGP000133
#> 6 2021-02-03 PGP000134
#> 7 2021-02-03 PGP000135
#> 8 2021-02-03 PGP000136
#> 

# Get all releases
get_releases(date = 'all')

#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 23 x 5
#>    date       n_pgs n_ppm n_pgp notes                                           
#>    <date>     <int> <int> <int> <chr>                                           
#>  1 2021-02-03    58   265     8 This release contains 58 new Score(s), 8 new Pu…
#>  2 2021-01-07     6    31     6 This release contains 6 new Score(s), 6 new Pub…
#>  3 2020-12-15   306   313     4 This release contains 306 new Score(s), 4 new P…
#>  4 2020-12-08     9    65     6 This release contains 9 new Score(s), 6 new Pub…
#>  5 2020-11-20     4    50     5 This release contains 4 new Score(s), 5 new Pub…
#>  6 2020-11-05     4    19     4 This release contains 4 new Score(s), 4 new Pub…
#>  7 2020-10-19    79    79     1 This release contains 79 new Score(s), 1 new Pu…
#>  8 2020-10-16     1     1     1 This release contains 1 new Score(s), 1 new Pub…
#>  9 2020-09-18    10    34     3 This release contains 10 new Score(s), 3 new Pu…
#> 10 2020-09-04     6    17     3 This release contains 6 new Score(s), 3 new Pub…
#> # … with 13 more rows
#> 
#> Slot "pgs_ids":
#> # A tibble: 721 x 2
#>    date       pgs_id   
#>    <date>     <chr>    
#>  1 2021-02-03 PGS000668
#>  2 2021-02-03 PGS000669
#>  3 2021-02-03 PGS000670
#>  4 2021-02-03 PGS000671
#>  5 2021-02-03 PGS000672
#>  6 2021-02-03 PGS000673
#>  7 2021-02-03 PGS000674
#>  8 2021-02-03 PGS000675
#>  9 2021-02-03 PGS000676
#> 10 2021-02-03 PGS000677
#> # … with 711 more rows
#> 
#> Slot "ppm_ids":
#> # A tibble: 1,533 x 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2021-02-03 PPM001396
#>  2 2021-02-03 PPM001397
#>  3 2021-02-03 PPM001398
#>  4 2021-02-03 PPM001399
#>  5 2021-02-03 PPM001400
#>  6 2021-02-03 PPM001401
#>  7 2021-02-03 PPM001402
#>  8 2021-02-03 PPM001403
#>  9 2021-02-03 PPM001404
#> 10 2021-02-03 PPM001405
#> # … with 1,523 more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 133 x 2
#>    date       pgp_id   
#>    <date>     <chr>    
#>  1 2021-02-03 PGP000128
#>  2 2021-02-03 PGP000129
#>  3 2021-02-03 PGP000130
#>  4 2021-02-03 PGP000132
#>  5 2021-02-03 PGP000133
#>  6 2021-02-03 PGP000134
#>  7 2021-02-03 PGP000135
#>  8 2021-02-03 PGP000136
#>  9 2021-01-07 PGP000122
#> 10 2021-01-07 PGP000123
#> # … with 123 more rows
#> 

# Get a specific release by date
get_releases(date = '2020-08-19')

#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 x 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2020-08-19    22    67     5 This release contains 22 new Score(s), 5 new Pub…
#> 
#> Slot "pgs_ids":
#> # A tibble: 22 x 2
#>    date       pgs_id   
#>    <date>     <chr>    
#>  1 2020-08-19 PGS000296
#>  2 2020-08-19 PGS000297
#>  3 2020-08-19 PGS000298
#>  4 2020-08-19 PGS000299
#>  5 2020-08-19 PGS000300
#>  6 2020-08-19 PGS000301
#>  7 2020-08-19 PGS000302
#>  8 2020-08-19 PGS000303
#>  9 2020-08-19 PGS000304
#> 10 2020-08-19 PGS000305
#> # … with 12 more rows
#> 
#> Slot "ppm_ids":
#> # A tibble: 67 x 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2020-08-19 PPM000743
#>  2 2020-08-19 PPM000744
#>  3 2020-08-19 PPM000745
#>  4 2020-08-19 PPM000746
#>  5 2020-08-19 PPM000747
#>  6 2020-08-19 PPM000748
#>  7 2020-08-19 PPM000749
#>  8 2020-08-19 PPM000750
#>  9 2020-08-19 PPM000751
#> 10 2020-08-19 PPM000752
#> # … with 57 more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 5 x 2
#>   date       pgp_id   
#>   <date>     <chr>    
#> 1 2020-08-19 PGP000090
#> 2 2020-08-19 PGP000091
#> 3 2020-08-19 PGP000092
#> 4 2020-08-19 PGP000093
#> 5 2020-08-19 PGP000094
#>