Get PGS Catalog Releases

This function retrieves PGS Catalog release information. Note that the columns pgs_id, ppm_id and pgp_id contain in each element a vector. These columns can be unnested using unnest_longer (see Examples).

get_releases(
  date = "latest",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

date
One or more dates formatted as "YYYY-MM-DD" for respective releases, "latest" for the latest release, or "all" for all releases.

verbose
Whether to print information about the underlying requests to the REST API server.

warnings
Whether to print warnings about the underlying requests to the REST API server.

progress_bar
Whether to show a progress bar indicating download progress from the REST API server.

Value

A data frame where each row is a release. Columns are:

date

Release date.

n_pgs

Number of released Polygenic Score (PGS) identifiers (pgs_id).

n_ppm

Number of released Performance Metric (PPM) identifiers (ppm_id).

n_pgp

Number of released PGS Catalog Publication (PGP) identifiers (pgp_id).

pgs_id

Released Polygenic Score (PGS) identifiers.

ppm_id

Released Performance Metric (PPM) identifiers.

pgp_id

Released PGS Catalog Publication (PGP) identifiers.

notes

News about the release.

Examples

# Get the latest release
get_releases()

#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 × 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2021-09-17     9    46     7 This release contains 9 new Score(s), 7 new Publ…
#> 
#> Slot "pgs_ids":
#> # A tibble: 9 × 2
#>   date       pgs_id   
#>   <date>     <chr>    
#> 1 2021-09-17 PGS000898
#> 2 2021-09-17 PGS000899
#> 3 2021-09-17 PGS000900
#> 4 2021-09-17 PGS000901
#> 5 2021-09-17 PGS000902
#> 6 2021-09-17 PGS000903
#> 7 2021-09-17 PGS000904
#> 8 2021-09-17 PGS000905
#> 9 2021-09-17 PGS000906
#> 
#> Slot "ppm_ids":
#> # A tibble: 46 × 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2021-09-17 PPM002634
#>  2 2021-09-17 PPM002635
#>  3 2021-09-17 PPM002636
#>  4 2021-09-17 PPM002637
#>  5 2021-09-17 PPM002638
#>  6 2021-09-17 PPM002639
#>  7 2021-09-17 PPM002640
#>  8 2021-09-17 PPM002641
#>  9 2021-09-17 PPM002642
#> 10 2021-09-17 PPM002643
#> # … with 36 more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 7 × 2
#>   date       pgp_id   
#>   <date>     <chr>    
#> 1 2021-09-17 PGP000231
#> 2 2021-09-17 PGP000232
#> 3 2021-09-17 PGP000233
#> 4 2021-09-17 PGP000234
#> 5 2021-09-17 PGP000235
#> 6 2021-09-17 PGP000236
#> 7 2021-09-17 PGP000237
#> 

get_releases(date = 'latest')

#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 × 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2021-09-17     9    46     7 This release contains 9 new Score(s), 7 new Publ…
#> 
#> Slot "pgs_ids":
#> # A tibble: 9 × 2
#>   date       pgs_id   
#>   <date>     <chr>    
#> 1 2021-09-17 PGS000898
#> 2 2021-09-17 PGS000899
#> 3 2021-09-17 PGS000900
#> 4 2021-09-17 PGS000901
#> 5 2021-09-17 PGS000902
#> 6 2021-09-17 PGS000903
#> 7 2021-09-17 PGS000904
#> 8 2021-09-17 PGS000905
#> 9 2021-09-17 PGS000906
#> 
#> Slot "ppm_ids":
#> # A tibble: 46 × 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2021-09-17 PPM002634
#>  2 2021-09-17 PPM002635
#>  3 2021-09-17 PPM002636
#>  4 2021-09-17 PPM002637
#>  5 2021-09-17 PPM002638
#>  6 2021-09-17 PPM002639
#>  7 2021-09-17 PPM002640
#>  8 2021-09-17 PPM002641
#>  9 2021-09-17 PPM002642
#> 10 2021-09-17 PPM002643
#> # … with 36 more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 7 × 2
#>   date       pgp_id   
#>   <date>     <chr>    
#> 1 2021-09-17 PGP000231
#> 2 2021-09-17 PGP000232
#> 3 2021-09-17 PGP000233
#> 4 2021-09-17 PGP000234
#> 5 2021-09-17 PGP000235
#> 6 2021-09-17 PGP000236
#> 7 2021-09-17 PGP000237
#> 

# Get all releases
get_releases(date = 'all')

#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 33 × 5
#>    date       n_pgs n_ppm n_pgp notes                                           
#>    <date>     <int> <int> <int> <chr>                                           
#>  1 2021-09-17     9    46     7 This release contains 9 new Score(s), 7 new Pub…
#>  2 2021-08-26    18   120    17 This release contains 18 new Score(s), 17 new P…
#>  3 2021-07-29    17    66    10 This release contains 17 new Score(s), 10 new P…
#>  4 2021-07-02     6    35     7 This release contains 6 new Score(s), 7 new Pub…
#>  5 2021-06-11    13    63     6 This release contains 13 new Score(s), 6 new Pu…
#>  6 2021-05-28    33   115    17 This release contains 33 new Score(s), 17 new P…
#>  7 2021-04-28     4    25     7 This release contains 4 new Score(s), 7 new Pub…
#>  8 2021-04-07     6    22     5 This release contains 6 new Score(s), 5 new Pub…
#>  9 2021-03-22    13   144    11 This release contains 13 new Score(s), 11 new P…
#> 10 2021-02-23    17   118    11 This release contains 17 new Score(s), 11 new P…
#> # … with 23 more rows
#> 
#> Slot "pgs_ids":
#> # A tibble: 856 × 2
#>    date       pgs_id   
#>    <date>     <chr>    
#>  1 2021-09-17 PGS000898
#>  2 2021-09-17 PGS000899
#>  3 2021-09-17 PGS000900
#>  4 2021-09-17 PGS000901
#>  5 2021-09-17 PGS000902
#>  6 2021-09-17 PGS000903
#>  7 2021-09-17 PGS000904
#>  8 2021-09-17 PGS000905
#>  9 2021-09-17 PGS000906
#> 10 2021-08-26 PGS000865
#> # … with 846 more rows
#> 
#> Slot "ppm_ids":
#> # A tibble: 2,286 × 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2021-09-17 PPM002634
#>  2 2021-09-17 PPM002635
#>  3 2021-09-17 PPM002636
#>  4 2021-09-17 PPM002637
#>  5 2021-09-17 PPM002638
#>  6 2021-09-17 PPM002639
#>  7 2021-09-17 PPM002640
#>  8 2021-09-17 PPM002641
#>  9 2021-09-17 PPM002642
#> 10 2021-09-17 PPM002643
#> # … with 2,276 more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 231 × 2
#>    date       pgp_id   
#>    <date>     <chr>    
#>  1 2021-09-17 PGP000231
#>  2 2021-09-17 PGP000232
#>  3 2021-09-17 PGP000233
#>  4 2021-09-17 PGP000234
#>  5 2021-09-17 PGP000235
#>  6 2021-09-17 PGP000236
#>  7 2021-09-17 PGP000237
#>  8 2021-08-26 PGP000212
#>  9 2021-08-26 PGP000213
#> 10 2021-08-26 PGP000214
#> # … with 221 more rows
#> 

# Get a specific release by date
get_releases(date = '2020-08-19')

#> An object of class "releases"
#> Slot "releases":
#> # A tibble: 1 × 5
#>   date       n_pgs n_ppm n_pgp notes                                            
#>   <date>     <int> <int> <int> <chr>                                            
#> 1 2020-08-19    22    67     5 This release contains 22 new Score(s), 5 new Pub…
#> 
#> Slot "pgs_ids":
#> # A tibble: 22 × 2
#>    date       pgs_id   
#>    <date>     <chr>    
#>  1 2020-08-19 PGS000296
#>  2 2020-08-19 PGS000297
#>  3 2020-08-19 PGS000298
#>  4 2020-08-19 PGS000299
#>  5 2020-08-19 PGS000300
#>  6 2020-08-19 PGS000301
#>  7 2020-08-19 PGS000302
#>  8 2020-08-19 PGS000303
#>  9 2020-08-19 PGS000304
#> 10 2020-08-19 PGS000305
#> # … with 12 more rows
#> 
#> Slot "ppm_ids":
#> # A tibble: 67 × 2
#>    date       ppm_id   
#>    <date>     <chr>    
#>  1 2020-08-19 PPM000743
#>  2 2020-08-19 PPM000744
#>  3 2020-08-19 PPM000745
#>  4 2020-08-19 PPM000746
#>  5 2020-08-19 PPM000747
#>  6 2020-08-19 PPM000748
#>  7 2020-08-19 PPM000749
#>  8 2020-08-19 PPM000750
#>  9 2020-08-19 PPM000751
#> 10 2020-08-19 PPM000752
#> # … with 57 more rows
#> 
#> Slot "pgp_ids":
#> # A tibble: 5 × 2
#>   date       pgp_id   
#>   <date>     <chr>    
#> 1 2020-08-19 PGP000090
#> 2 2020-08-19 PGP000091
#> 3 2020-08-19 PGP000092
#> 4 2020-08-19 PGP000093
#> 5 2020-08-19 PGP000094
#>