Get PGS Catalog Sample Sets

Retrieves sample sets via the PGS Catalog REST API. The REST API is queried multiple times with the criteria passed as arguments (see below). By default all sample sets that match the criteria supplied in the arguments are retrieved: this corresponds to the default option set_operation set to 'union'. If you rather have only the associations that match simultaneously all criteria provided, then set set_operation to 'intersection'.

get_sample_sets(
  pss_id = NULL,
  pgs_id = NULL,
  set_operation = "union",
  interactive = TRUE,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

pss_id
A character vector of PGS Catalog sample sets accession identifiers.

pgs_id
A character vector of PGS Catalog score accession identifiers.

set_operation
Either 'union' or 'intersection'. This tells how performance metrics retrieved by different criteria should be combined: 'union' binds together all results removing duplicates and 'intersection' only keeps same sample sets found with different criteria.

interactive
A logical. If all sample sets are requested, whether to ask interactively if we really want to proceed.

verbose
A logical indicating whether the function should be verbose about the different queries or not.

warnings
A logical indicating whether to print warnings, if any.

progress_bar
Whether to show a progress bar indicating download progress from the REST API server.

Value

A sample_sets object.

Details

Please note that all search criteria are vectorised, thus allowing for batch mode search.

Examples

# Search by PGS identifier
get_sample_sets(pgs_id = 'PGS000013')

#> An object of class "sample_sets"
#> Slot "sample_sets":
#> # A tibble: 41 × 1
#>    pss_id   
#>    <chr>    
#>  1 PSS000015
#>  2 PSS000019
#>  3 PSS000020
#>  4 PSS000021
#>  5 PSS000022
#>  6 PSS000219
#>  7 PSS000227
#>  8 PSS000228
#>  9 PSS000229
#> 10 PSS000230
#> # … with 31 more rows
#> 
#> Slot "samples":
#> # A tibble: 45 × 15
#>    pss_id    sample_id stage sample_size sample_cases sample_controls
#>    <chr>         <int> <chr>       <int>        <int>           <int>
#>  1 PSS000015         1 eval       288978         8676          280302
#>  2 PSS000019         1 eval         5762          173            5589
#>  3 PSS000020         1 eval          862          446             416
#>  4 PSS000020         2 eval         2333          937            1396
#>  5 PSS000021         1 eval         1964          974             976
#>  6 PSS000022         1 eval         3309         2492             817
#>  7 PSS000219         1 eval        11010          126           10884
#>  8 PSS000227         1 eval          544           40             504
#>  9 PSS000228         1 eval         1298          336             962
#> 10 PSS000229         1 eval          919          168             751
#> # … with 35 more rows, and 9 more variables: sample_percent_male <dbl>,
#> #   phenotype_description <chr>, ancestry_category <chr>, ancestry <chr>,
#> #   country <chr>, ancestry_additional_description <chr>, study_id <chr>,
#> #   pubmed_id <chr>, cohorts_additional_description <chr>
#> 
#> Slot "demographics":
#> # A tibble: 16 × 11
#>    pss_id    sample_id variable       estimate_type estimate unit  variability_type
#>    <chr>         <int> <chr>          <chr>            <dbl> <chr> <chr>           
#>  1 PSS000365         1 age            mean              34   years NA              
#>  2 PSS000365         2 age            mean              33   years NA              
#>  3 PSS000366         1 age            mean              54   years NA              
#>  4 PSS000366         2 age            mean              55   years NA              
#>  5 PSS000367         1 age            mean              60.6 years NA              
#>  6 PSS000367         2 age            mean              52.8 years NA              
#>  7 PSS000331         1 follow_up_time median             9.2 years NA              
#>  8 PSS000332         1 follow_up_time median             9.2 years NA              
#>  9 PSS000333         1 follow_up_time median            11.7 years NA              
#> 10 PSS000334         1 follow_up_time median            11.7 years NA              
#> 11 PSS000335         1 follow_up_time median            10.4 years NA              
#> 12 PSS000336         1 follow_up_time median            10.4 years NA              
#> 13 PSS000467         1 follow_up_time median            21.3 years NA              
#> 14 PSS000468         1 follow_up_time median            23.2 years NA              
#> 15 PSS000469         1 follow_up_time median             8.1 years NA              
#> 16 PSS001063         1 follow_up_time median            14   years NA              
#> # … with 4 more variables: variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> 
#> Slot "cohorts":
#> # A tibble: 115 × 4
#>    pss_id    sample_id cohort_symbol cohort_name                                
#>    <chr>         <int> <chr>         <chr>                                      
#>  1 PSS000015         1 UKB           UK Biobank                                 
#>  2 PSS000019         1 CARTaGENE     CARTaGENE cohort (CHU Sainte-Justine, Queb…
#>  3 PSS000020         1 MHI           Montreal Heart Institute Biobank           
#>  4 PSS000020         2 MHI           Montreal Heart Institute Biobank           
#>  5 PSS000021         1 MHI           Montreal Heart Institute Biobank           
#>  6 PSS000022         1 MHI           Montreal Heart Institute Biobank           
#>  7 PSS000219         1 CG            Color Genomics                             
#>  8 PSS000227         1 VIRGO         Variation in Recovery: Role of Gender on O…
#>  9 PSS000227         1 MESA          Multi-Ethnic Study of Atherosclerosis      
#> 10 PSS000228         1 VIRGO         Variation in Recovery: Role of Gender on O…
#> # … with 105 more rows
#> 

# Search by the PSS identifier
get_sample_sets(pss_id = 'PSS000068')

#> An object of class "sample_sets"
#> Slot "sample_sets":
#> # A tibble: 1 × 1
#>   pss_id   
#>   <chr>    
#> 1 PSS000068
#> 
#> Slot "samples":
#> # A tibble: 3 × 15
#>   pss_id    sample_id stage sample_size sample_cases sample_controls
#>   <chr>         <int> <chr>       <int>        <int>           <int>
#> 1 PSS000068         1 eval         3157         3157               0
#> 2 PSS000068         2 eval         8796         2148            6648
#> 3 PSS000068         3 eval         5059         4099             960
#> # … with 9 more variables: sample_percent_male <dbl>,
#> #   phenotype_description <chr>, ancestry_category <chr>, ancestry <chr>,
#> #   country <chr>, ancestry_additional_description <chr>, study_id <chr>,
#> #   pubmed_id <chr>, cohorts_additional_description <chr>
#> 
#> Slot "demographics":
#> # A tibble: 3 × 11
#>   pss_id    sample_id variable estimate_type estimate unit  variability_type
#>   <chr>         <int> <chr>    <chr>            <dbl> <chr> <chr>           
#> 1 PSS000068         1 age      NA                  NA years NA              
#> 2 PSS000068         2 age      NA                  NA years NA              
#> 3 PSS000068         3 age      NA                  NA years NA              
#> # … with 4 more variables: variability <dbl>, interval_type <chr>,
#> #   interval_lower <dbl>, interval_upper <dbl>
#> 
#> Slot "cohorts":
#> # A tibble: 3 × 4
#>   pss_id    sample_id cohort_symbol cohort_name                                 
#>   <chr>         <int> <chr>         <chr>                                       
#> 1 PSS000068         1 UKGPCS        UK Genetic Prostate Cancer Study            
#> 2 PSS000068         2 ProtecT       Prostate testing for cancer and Treatment s…
#> 3 PSS000068         3 SEARCH        Study of Epidemiology and Risk Factors in C…
#>