Read a polygenic scoring file

This function imports a PGS scoring file. For more information about the scoring file schema check vignette("pgs-scoring-file", package = "quincunx").

read_scoring_file(source, protocol = "http", metadata_only = FALSE)

Arguments

source
PGS scoring file. This can be specified in three forms: (i) a PGS identifier, e.g. "PGS000001", (ii) a path to a local file, e.g. "~/PGS000001.txt" or "~/PGS000001.txt.gz" or (iii) a direct URL to the PGS Catalog FTP server, e.g. "http://ftp.ebi.ac.uk/pub/databases/spot/pgs/scores/PGS000001/ScoringFiles/PGS000001.txt.gz".

protocol
Network protocol for communication with the PGS Catalog FTP server: either "http" or "ftp".

metadata_only
Whether to read only the comment block (header) from the scoring file.

Value

The returned value is a named list. The names are copied from the arguments passed in source. Each element of the list contains another list of two elements: "metadata" and "data". The "metadata" element contains data parsed from the header of the PGS scoring file. The "data" element contains a data frame with as many rows as variants that constitute the PGS score. The columns can vary. There are mandatory and optional columns. The mandatory columns are those that identify the variant, effect allele (effect_allele), and its respective weight (effect_weight) in the score. The columns that identify the variant can either be the rsID or the combination of chr_name and chr_position. The "data" element will be NULL is argument metadata_only is TRUE. For more information about the scoring file schema check vignette("pgs-scoring-file", package = "quincunx").

Examples

# Read a PGS scoring file by PGS ID
# (internally, it translates the PGS ID
#  to the corresponding FTP URL)
read_scoring_file("PGS000655")

#> $PGS000655
#> $PGS000655$metadata
#> # A tibble: 1 x 7
#>   pgs_id reported_trait original_genome… number_of_varia… pgp_id citation
#>   <chr>  <chr>          <chr>                       <int> <chr>  <chr>   
#> 1 PGS00… Non-alcoholic… GRCh37                         10 PGP00… Namjou …
#> # … with 1 more variable: license <chr>
#> 
#> $PGS000655$data
#> # A tibble: 10 x 5
#>    rsID       effect_allele effect_weight weight_type    OR
#>    <chr>      <chr>                 <dbl> <chr>       <dbl>
#>  1 rs1260326  T                    0.0571 beta         1.14
#>  2 rs780094   T                    0.0611 beta         1.15
#>  3 rs2954021  G                   -0.0980 beta         0.8 
#>  4 rs4808199  A                    0.0868 beta         1.22
#>  5 rs58542926 T                    0.0899 beta         1.23
#>  6 rs738409   G                    0.253  beta         1.79
#>  7 rs3747207  A                    0.250  beta         1.78
#>  8 rs2294915  T                    0.243  beta         1.75
#>  9 rs1227756  G                    0.0261 beta         1.06
#> 10 rs2294918  A                   -0.114  beta         0.78
#> 
#> 

# Equivalent to `read_scoring_file("PGS000655")`
url <- paste0(
  "http://ftp.ebi.ac.uk/",
  "pub/databases/spot/pgs/scores/",
  "PGS000655/ScoringFiles/",
  "PGS000655.txt.gz"
)
read_scoring_file(url)

#> $`http://ftp.ebi.ac.uk/pub/databases/spot/pgs/scores/PGS000655/ScoringFiles/PGS000655.txt.gz`
#> $`http://ftp.ebi.ac.uk/pub/databases/spot/pgs/scores/PGS000655/ScoringFiles/PGS000655.txt.gz`$metadata
#> # A tibble: 1 x 7
#>   pgs_id reported_trait original_genome… number_of_varia… pgp_id citation
#>   <chr>  <chr>          <chr>                       <int> <chr>  <chr>   
#> 1 PGS00… Non-alcoholic… GRCh37                         10 PGP00… Namjou …
#> # … with 1 more variable: license <chr>
#> 
#> $`http://ftp.ebi.ac.uk/pub/databases/spot/pgs/scores/PGS000655/ScoringFiles/PGS000655.txt.gz`$data
#> # A tibble: 10 x 5
#>    rsID       effect_allele effect_weight weight_type    OR
#>    <chr>      <chr>                 <dbl> <chr>       <dbl>
#>  1 rs1260326  T                    0.0571 beta         1.14
#>  2 rs780094   T                    0.0611 beta         1.15
#>  3 rs2954021  G                   -0.0980 beta         0.8 
#>  4 rs4808199  A                    0.0868 beta         1.22
#>  5 rs58542926 T                    0.0899 beta         1.23
#>  6 rs738409   G                    0.253  beta         1.79
#>  7 rs3747207  A                    0.250  beta         1.78
#>  8 rs2294915  T                    0.243  beta         1.75
#>  9 rs1227756  G                    0.0261 beta         1.06
#> 10 rs2294918  A                   -0.114  beta         0.78
#> 
#> 

if (FALSE) {
# Reading from a local file
read_scoring_file("~/PGS000655.txt.gz")
}