An S4 class to represent a set of PGS Catalog Sample Sets

The sample_sets object consists of four slots, each a table (tibble), that combined form a relational database of a subset of PGS Catalog sample sets. Each sample set is an observation (row) in the sample_sets table --- main table. All tables have the column pss_id as primary key.

Slots

sample_sets

TODO

TODO

TODO

TODO

samples

A tibble listing TODO. Columns:

TODO

TODO

TODO

TODO

demographics

A tibble listing TODO. Columns:

TODO

TODO

TODO

TODO

cohorts

A tibble listing TODO. Columns:

TODO

TODO

TODO

TODO